% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_basic_variables}
\alias{add_basic_variables}
\title{Add Basic Variables}
\usage{
add_basic_variables(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional basic variables.
}
\description{
Add Basic Variables
}
\keyword{internal}
