% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_chain_variables}
\alias{add_chain_variables}
\title{Add Chain Variables}
\usage{
add_chain_variables(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional chain-related variables.
}
\description{
Add Chain Variables
}
\keyword{internal}
