% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{add_epv_variables}
\alias{add_epv_variables}
\title{Add Expected Points Value (EPV) Variables}
\usage{
add_epv_variables(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional EPV variables.
}
\description{
This function adds EPV-related variables to the dataframe.
}
\keyword{internal}
