% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_game_variables}
\alias{add_game_variables}
\title{Add Game Variables}
\usage{
add_game_variables(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional game-level variables.
}
\description{
Add Game Variables
}
\keyword{internal}
