% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{add_lagged_variables}
\alias{add_lagged_variables}
\title{Add Lagged Variables}
\usage{
add_lagged_variables(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional lagged variables.
}
\description{
Add Lagged Variables
}
\keyword{internal}
