% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_quarter_variables}
\alias{add_quarter_variables}
\title{Add Quarter Variables}
\usage{
add_quarter_variables(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional quarter-level variables.
}
\description{
Add Quarter Variables
}
\keyword{internal}
