% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_score_variables}
\alias{add_score_variables}
\title{Add Score Variables}
\usage{
add_score_variables(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional score-related variables.
}
\description{
Add Score Variables
}
\keyword{internal}
