% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{add_shot_result_variables}
\alias{add_shot_result_variables}
\title{Add Shot Result Variables}
\usage{
add_shot_result_variables(df)
}
\arguments{
\item{df}{A dataframe containing shot data.}
}
\value{
A dataframe with additional shot result variables.
}
\description{
Add Shot Result Variables
}
\keyword{internal}
