% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{add_shot_type_variables}
\alias{add_shot_type_variables}
\title{Add Shot Type Variables}
\usage{
add_shot_type_variables(df)
}
\arguments{
\item{df}{A dataframe containing shot data.}
}
\value{
A dataframe with additional shot type variables.
}
\description{
Add Shot Type Variables
}
\keyword{internal}
