% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_variables.R
\name{add_shot_vars}
\alias{add_shot_vars}
\title{Add Shot Variables}
\usage{
add_shot_vars(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional shot-related variables.
}
\description{
This function adds shot-related variables to the input dataframe.
}
