% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{add_speed_variables}
\alias{add_speed_variables}
\title{Add Speed Variables}
\usage{
add_speed_variables(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional speed variables.
}
\description{
Add Speed Variables
}
\keyword{internal}
