% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_torp_ids}
\alias{add_torp_ids}
\title{Add TORP IDs}
\usage{
add_torp_ids(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional TORP ID columns.
}
\description{
Add TORP IDs
}
\keyword{internal}
