% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_torp_ids_dt}
\alias{add_torp_ids_dt}
\title{Add TORP IDs and basic variables (Pass 1)}
\usage{
add_torp_ids_dt(dt)
}
\arguments{
\item{dt}{A data.table to modify}
}
\value{
Invisible NULL (modifies dt by reference)
}
\description{
Adds TORP identifiers and basic computed variables to the data.table.
Modifies the data.table by reference.
}
\keyword{internal}
