% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_API_BASE_URL}
\alias{AFL_API_BASE_URL}
\title{AFL public API base URL (v2)}
\format{
An object of class \code{character} of length 1.
}
\usage{
AFL_API_BASE_URL
}
\description{
AFL public API base URL (v2)
}
\keyword{internal}
