% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_CFS_API_BASE_URL}
\alias{AFL_CFS_API_BASE_URL}
\title{AFL CFS API base URL (fixtures, results, players)}
\format{
An object of class \code{character} of length 1.
}
\usage{
AFL_CFS_API_BASE_URL
}
\description{
AFL CFS API base URL (fixtures, results, players)
}
\keyword{internal}
