% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_COMP_PUBLIC_API_CODES}
\alias{AFL_COMP_PUBLIC_API_CODES}
\title{AFL public API (\code{aflapi.afl.com.au}) competition \code{code} values per comp}
\format{
An object of class \code{list} of length 2.
}
\usage{
AFL_COMP_PUBLIC_API_CODES
}
\description{
Used to filter the \code{competitions} endpoint response in
\code{\link[=.afl_all_comp_seasons]{.afl_all_comp_seasons()}}. AFLM matches multiple historical code spellings;
AFLW has always used a single one.
}
\keyword{internal}
