% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_DEFAULT_TIMEZONE}
\alias{AFL_DEFAULT_TIMEZONE}
\title{Default timezone for AFL matches}
\format{
An object of class \code{character} of length 1.
}
\usage{
AFL_DEFAULT_TIMEZONE
}
\description{
Default timezone for AFL matches
}
\keyword{internal}
