% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_GOAL_WIDTH}
\alias{AFL_GOAL_WIDTH}
\title{Width of an AFL goal in metres (between the goal posts)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFL_GOAL_WIDTH
}
\description{
Width of an AFL goal in metres (between the goal posts)
}
\keyword{internal}
