% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_MAX_PERIODS}
\alias{AFL_MAX_PERIODS}
\title{Number of regulation quarters in an AFL match}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFL_MAX_PERIODS
}
\description{
Number of regulation quarters in an AFL match
}
\keyword{internal}
