% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_MAX_REGULAR_ROUNDS}
\alias{AFL_MAX_REGULAR_ROUNDS}
\title{Maximum round number fallback when AFL_REGULAR_SEASON_ROUNDS lookup fails}
\format{
An object of class \code{integer} of length 1.
}
\usage{
AFL_MAX_REGULAR_ROUNDS
}
\description{
Maximum round number fallback when AFL_REGULAR_SEASON_ROUNDS lookup fails
}
\keyword{internal}
