% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_MIN_SEASON}
\alias{AFL_MIN_SEASON}
\title{Earliest season with data available in torpdata}
\format{
An object of class \code{integer} of length 1.
}
\usage{
AFL_MIN_SEASON
}
\description{
Earliest season with data available in torpdata
}
\keyword{internal}
