% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_PLAY_GAME_SECONDS}
\alias{AFL_PLAY_GAME_SECONDS}
\title{Nominal total playing time per AFL match in seconds (4 x 1200)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
AFL_PLAY_GAME_SECONDS
}
\description{
Nominal total playing time per AFL match in seconds (4 x 1200)
}
\keyword{internal}
