% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_REGULAR_SEASON_ROUNDS}
\alias{AFL_REGULAR_SEASON_ROUNDS}
\title{Regular season rounds by year (excludes finals)
2022 and earlier: 23 rounds; 2023 onwards: 24 rounds}
\format{
An object of class \code{integer} of length 6.
}
\usage{
AFL_REGULAR_SEASON_ROUNDS
}
\description{
Regular season rounds by year (excludes finals)
2022 and earlier: 23 rounds; 2023 onwards: 24 rounds
}
\keyword{internal}
