% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_ROOF_VENUES}
\alias{AFL_ROOF_VENUES}
\title{Venues with a closed roof (weather features set to neutral)}
\format{
An object of class \code{character} of length 1.
}
\usage{
AFL_ROOF_VENUES
}
\description{
Venues with a closed roof (weather features set to neutral)
}
\keyword{internal}
