% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_SAPI_BASE_URL}
\alias{AFL_SAPI_BASE_URL}
\title{AFL chain/play data API base URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
AFL_SAPI_BASE_URL
}
\description{
AFL chain/play data API base URL
}
\keyword{internal}
