% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_TEAM_SIZE}
\alias{AFL_TEAM_SIZE}
\title{Number of on-field players per team in a standard AFL match}
\format{
An object of class \code{integer} of length 1.
}
\usage{
AFL_TEAM_SIZE
}
\description{
Number of on-field players per team in a standard AFL match
}
\keyword{internal}
