% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_TIME_SCALER_MAX}
\alias{AFL_TIME_SCALER_MAX}
\title{Maximum value for time left scaler calculation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFL_TIME_SCALER_MAX
}
\description{
Maximum value for time left scaler calculation
}
\keyword{internal}
