% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_TOTAL_GAME_SECONDS}
\alias{AFL_TOTAL_GAME_SECONDS}
\title{Default total game seconds (4 quarters * 2000s)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFL_TOTAL_GAME_SECONDS
}
\description{
Default total game seconds (4 quarters * 2000s)
}
\keyword{internal}
