% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_aflw.R
\docType{data}
\name{AFLW_ELO_CARRYOVER}
\alias{AFLW_ELO_CARRYOVER}
\title{AFLW Elo season-boundary carryover (fraction of rating retained)
Seeded from men's ELO_CARRYOVER -- not yet retuned. See AFLW_ELO_K.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFLW_ELO_CARRYOVER
}
\description{
AFLW Elo season-boundary carryover (fraction of rating retained)
Seeded from men's ELO_CARRYOVER -- not yet retuned. See AFLW_ELO_K.
}
\keyword{internal}
