% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_aflw.R
\docType{data}
\name{AFLW_ELO_HGA}
\alias{AFLW_ELO_HGA}
\title{AFLW Elo home-ground advantage, in Elo points
Seeded from men's ELO_HGA -- not yet retuned. See AFLW_ELO_K.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFLW_ELO_HGA
}
\description{
AFLW Elo home-ground advantage, in Elo points
Seeded from men's ELO_HGA -- not yet retuned. See AFLW_ELO_K.
}
\keyword{internal}
