% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_aflw.R
\docType{data}
\name{AFLW_ELO_K}
\alias{AFLW_ELO_K}
\title{AFLW Elo update rate}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFLW_ELO_K
}
\description{
Seeded from men's ELO_K as a starting prior -- NOT yet retuned on an AFLW-specific
grid search. Re-tune once enough AFLW seasons of results exist to run the same
walk-forward backtest used for the men's constant, then update this comment.
}
\keyword{internal}
