% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contests.R
\name{build_h2h_summary}
\alias{build_h2h_summary}
\title{Build head-to-head summary table}
\usage{
build_h2h_summary(contests, p1, p2)
}
\arguments{
\item{contests}{data.table of filtered contests}

\item{p1}{Player 1 info list}

\item{p2}{Player 2 info list}
}
\value{
data.table summary
}
\description{
Build head-to-head summary table
}
\keyword{internal}
