% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_models.R
\name{calculate_brier_score}
\alias{calculate_brier_score}
\title{Simple Brier Score Calculation}
\usage{
calculate_brier_score(actual, predicted)
}
\arguments{
\item{actual}{Vector of actual binary outcomes}

\item{predicted}{Vector of predicted probabilities}
}
\value{
Numeric Brier score value
}
\description{
Simple Brier Score Calculation
}
\keyword{internal}
