% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_home_points_row}
\alias{calculate_home_points_row}
\title{Calculate Home Points Row}
\usage{
calculate_home_points_row(home, final_state, description, points_row)
}
\arguments{
\item{home}{Whether it's the home team (1) or away team (0).}

\item{final_state}{The final state of the play.}

\item{description}{The description of the play.}

\item{points_row}{The points for the row.}
}
\value{
A numeric value representing the home team's points for the row.
}
\description{
Calculate Home Points Row
}
\keyword{internal}
