% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_label_wp}
\alias{calculate_label_wp}
\title{Calculate Label Win Probability}
\usage{
calculate_label_wp(home_score, away_score, home)
}
\arguments{
\item{home_score}{The home team's score.}

\item{away_score}{The away team's score.}

\item{home}{Whether it's the home team (1) or away team (0).}
}
\value{
A numeric value representing the win probability label.
}
\description{
Calculate Label Win Probability
}
\keyword{internal}
