% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{calculate_lagged_coordinate}
\alias{calculate_lagged_coordinate}
\title{Calculate Lagged Coordinate}
\usage{
calculate_lagged_coordinate(coord, team_id_mdl)
}
\arguments{
\item{coord}{A vector of coordinates.}

\item{team_id_mdl}{A vector of team IDs for modeling.}
}
\value{
A vector of lagged coordinates.
}
\description{
Calculate Lagged Coordinate
}
\keyword{internal}
