% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{calculate_lagged_goal_x}
\alias{calculate_lagged_goal_x}
\title{Calculate Lagged Goal X}
\usage{
calculate_lagged_goal_x(goal_x, team_id_mdl, venue_length, lag = 1)
}
\arguments{
\item{goal_x}{A vector of goal x-coordinates.}

\item{team_id_mdl}{A vector of team IDs for modeling.}

\item{venue_length}{The length of the venue.}

\item{lag}{The number of positions to lag (default is 1).}
}
\value{
A vector of lagged goal x-coordinates.
}
\description{
Calculate Lagged Goal X
}
\keyword{internal}
