% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_models.R
\name{calculate_log_loss}
\alias{calculate_log_loss}
\title{Simple Log Loss Calculation}
\usage{
calculate_log_loss(actual, predicted)
}
\arguments{
\item{actual}{Vector of actual binary outcomes}

\item{predicted}{Vector of predicted probabilities}
}
\value{
Numeric log loss value
}
\description{
Simple Log Loss Calculation
}
\keyword{internal}
