% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{calculate_mirror}
\alias{calculate_mirror}
\title{Calculate Mirror}
\usage{
calculate_mirror(throw_in, team_id_mdl, x)
}
\arguments{
\item{throw_in}{A vector indicating if the play is a throw-in.}

\item{team_id_mdl}{A vector of team IDs for modeling.}

\item{x}{A vector of x-coordinates.}
}
\value{
A vector of mirror values (-1 or 1).
}
\description{
Calculate Mirror
}
\keyword{internal}
