% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{calculate_player_stats}
\alias{calculate_player_stats}
\title{Calculate player statistics}
\usage{
calculate_player_stats(
  player_game_data = NULL,
  match_ref,
  date_val,
  decay,
  loading,
  prior_games_recv,
  prior_games_disp,
  prior_games_spoil = RATING_PRIOR_GAMES_SPOIL,
  prior_games_hitout = RATING_PRIOR_GAMES_HITOUT
)
}
\arguments{
\item{player_game_data}{Player game data frame}

\item{match_ref}{Match reference}

\item{date_val}{Date value}

\item{decay}{Decay factor}

\item{loading}{Loading factor}

\item{prior_games_recv}{Prior games for receiving}

\item{prior_games_disp}{Prior games for disposal}

\item{prior_games_spoil}{Prior games for spoil shrinkage. Default is \code{RATING_PRIOR_GAMES_SPOIL}.}

\item{prior_games_hitout}{Prior games for hitout shrinkage. Default is \code{RATING_PRIOR_GAMES_HITOUT}.}
}
\value{
A data frame with calculated player statistics
}
\description{
Calculate player statistics
}
