% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_points_id}
\alias{calculate_points_id}
\title{Calculate Home Points Row}
\usage{
calculate_points_id(
  final_state,
  description,
  end_of_chain,
  end_of_qtr,
  team_id,
  opp_id
)
}
\arguments{
\item{final_state}{The final state of the play.}

\item{description}{The description of the play.}

\item{home}{Whether it's the home team (1) or away team (0).}

\item{points_row}{The points for the row.}
}
\value{
A numeric value representing the home team's points for the row.
}
\description{
Calculate Home Points Row
}
\keyword{internal}
