% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_points_row}
\alias{calculate_points_row}
\title{Calculate Points Row}
\usage{
calculate_points_row(final_state, description, end_of_chain, end_of_qtr)
}
\arguments{
\item{final_state}{The final state of the play.}

\item{description}{The description of the play.}
}
\value{
A numeric value representing the points for the row.
}
\description{
Calculate Points Row
}
\keyword{internal}
