% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{calculate_pos_lead_prob}
\alias{calculate_pos_lead_prob}
\title{Calculate Positive Lead Probability}
\usage{
calculate_pos_lead_prob(
  points_diff,
  opp_goal,
  opp_behind,
  no_score,
  behind,
  goal
)
}
\arguments{
\item{points_diff}{The points difference.}

\item{opp_goal}{Opponent's goal probability.}

\item{opp_behind}{Opponent's behind probability.}

\item{no_score}{No score probability.}

\item{behind}{Behind probability.}

\item{goal}{Goal probability.}
}
\value{
A numeric value representing the positive lead probability.
}
\description{
Calculate Positive Lead Probability
}
\keyword{internal}
