% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{calculate_quality_score}
\alias{calculate_quality_score}
\title{Calculate Overall Data Quality Score}
\usage{
calculate_quality_score(data, issues)
}
\arguments{
\item{data}{Dataframe being assessed}

\item{issues}{List of quality issues found}
}
\value{
Numeric quality score between 0 and 1
}
\description{
Calculates a composite quality score based on various quality metrics
}
\keyword{internal}
