% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_time_remaining}
\alias{calculate_time_remaining}
\title{Calculate time remaining in an AFL match (seconds)}
\usage{
calculate_time_remaining(period, period_seconds)
}
\arguments{
\item{period}{Numeric vector of period numbers (1-4)}

\item{period_seconds}{Numeric vector of seconds elapsed in the current period}
}
\value{
Numeric vector of seconds remaining in the match
}
\description{
Calculate time remaining in an AFL match (seconds)
}
\keyword{internal}
