% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_time_remaining_pct}
\alias{calculate_time_remaining_pct}
\title{Calculate time remaining as a proportion of total game time}
\usage{
calculate_time_remaining_pct(period, period_seconds)
}
\arguments{
\item{period}{Numeric vector of period numbers (1-4)}

\item{period_seconds}{Numeric vector of seconds elapsed in the current period}
}
\value{
Numeric vector of proportions (0 to 1)
}
\description{
Calculate time remaining as a proportion of total game time
}
\keyword{internal}
