% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{calculate_torp_ratings}
\alias{calculate_torp_ratings}
\title{Deprecated: use \code{\link{calculate_epr}} or \code{\link{torp_ratings}}}
\usage{
calculate_torp_ratings(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{calculate_epr}}.}
}
\value{
See \code{\link{calculate_epr}}.
}
\description{
Deprecated: use \code{\link{calculate_epr}} or \code{\link{torp_ratings}}
}
\keyword{internal}
