% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHAINS_MARK_WIN_DESCS}
\alias{CHAINS_MARK_WIN_DESCS}
\title{Chains descriptions indicating the opponent won the mark}
\format{
An object of class \code{character} of length 3.
}
\usage{
CHAINS_MARK_WIN_DESCS
}
\description{
Chains descriptions indicating the opponent won the mark
}
\keyword{internal}
