% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{check_internet_connection}
\alias{check_internet_connection}
\title{Check if internet connection is available}
\usage{
check_internet_connection()
}
\value{
Logical indicating if internet connection is available
}
\description{
Check if internet connection is available
}
\keyword{internal}
