% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_utils.R
\name{check_wp_model_health}
\alias{check_wp_model_health}
\title{Check Win Probability Model Health}
\usage{
check_wp_model_health()
}
\value{
List with health check results
}
\description{
This function is intended for internal use and may be unexported in a future release.
Performs basic health checks on the win probability model
}
