% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{clean_model_data_wp}
\alias{clean_model_data_wp}
\title{Clean Model Data for Win Probability (WP)}
\usage{
clean_model_data_wp(df)
}
\arguments{
\item{df}{A dataframe containing raw play-by-play data.}
}
\value{
A cleaned dataframe ready for WP modeling.
}
\description{
Clean Model Data for Win Probability (WP)
}
