% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{clean_shots_data}
\alias{clean_shots_data}
\title{Clean Shots Data}
\usage{
clean_shots_data(df)
}
\arguments{
\item{df}{A dataframe containing raw shot data.}
}
\value{
A cleaned dataframe with additional shot-related variables.
}
\description{
Clean Shots Data
}
