% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{CLOCK_DELTA_CAP}
\alias{CLOCK_DELTA_CAP}
\title{Maximum reasonable seconds between consecutive in-play chain events
Used to cap clock deltas when reconstructing play time.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
CLOCK_DELTA_CAP
}
\description{
Maximum reasonable seconds between consecutive in-play chain events
Used to cap clock deltas when reconstructing play time.
}
\keyword{internal}
