% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CLOCK_RESTART_EVENTS}
\alias{CLOCK_RESTART_EVENTS}
\title{Descriptions that always follow dead time (restarts after clock stoppage)}
\format{
An object of class \code{character} of length 9.
}
\usage{
CLOCK_RESTART_EVENTS
}
\description{
Descriptions that always follow dead time (restarts after clock stoppage)
}
\keyword{internal}
