% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{CLOCK_STOPPAGE_TRIGGERS}
\alias{CLOCK_STOPPAGE_TRIGGERS}
\title{Events that stop the play clock (chain \code{description} values)}
\format{
An object of class \code{character} of length 6.
}
\usage{
CLOCK_STOPPAGE_TRIGGERS
}
\description{
Events that stop the play clock (chain \code{description} values)
}
\keyword{internal}
