% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{compute_centrality_scores}
\alias{compute_centrality_scores}
\title{Compute centrality via Power Iteration}
\usage{
compute_centrality_scores(adj, damping = 0.85, max_iter = 100L, tol = 1e-06)
}
\description{
Compute centrality via Power Iteration
}
\keyword{internal}
