% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_credit.R
\name{compute_failed_recv_credit}
\alias{compute_failed_recv_credit}
\title{Backward-compatible wrapper (deprecated)}
\usage{
compute_failed_recv_credit(
  chains,
  weight_per_loss = EPV_RECV_FAILED_CONTEST_WT
)
}
\description{
Backward-compatible wrapper (deprecated)
}
\keyword{internal}
