% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CONTEST_GROUND_BALL_DESCS}
\alias{CONTEST_GROUND_BALL_DESCS}
\title{Descriptions for ground ball contests (present in both chains and PBP)
Adjacent ground ball rows from opposing teams at same x,y indicate a contest}
\format{
An object of class \code{character} of length 5.
}
\usage{
CONTEST_GROUND_BALL_DESCS
}
\description{
Descriptions for ground ball contests (present in both chains and PBP)
Adjacent ground ball rows from opposing teams at same x,y indicate a contest
}
\keyword{internal}
