% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_BEHINDS_WT}
\alias{CREDIT_BEHINDS_WT}
\title{Behinds weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_BEHINDS_WT
}
\description{
Behinds weight (disp component)
}
\keyword{internal}
