% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_BOUNCE_PENALTY}
\alias{CREDIT_BOUNCE_PENALTY}
\title{Bounce penalty per bounce}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_BOUNCE_PENALTY
}
\description{
Bounce penalty per bounce
}
\keyword{internal}
