% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_CLANGERS_WT}
\alias{CREDIT_CLANGERS_WT}
\title{Clangers weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_CLANGERS_WT
}
\description{
Clangers weight (disp component)
}
\keyword{internal}
