% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_CONTESTED_MARKS_WT}
\alias{CREDIT_CONTESTED_MARKS_WT}
\title{Contested marks weight (recv component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_CONTESTED_MARKS_WT
}
\description{
Contested marks weight (recv component)
}
\keyword{internal}
