% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_DEF_PRESSURE_WT}
\alias{CREDIT_DEF_PRESSURE_WT}
\title{Defensive half pressure act weight (spoil component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_DEF_PRESSURE_WT
}
\description{
Defensive half pressure act weight (spoil component)
}
\keyword{internal}
