% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_DISP_POS_OFFSET}
\alias{CREDIT_DISP_POS_OFFSET}
\title{Disposal EPV offset when possessing (pos_team == 1)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_DISP_POS_OFFSET
}
\description{
Disposal EPV offset when possessing (pos_team == 1)
}
\keyword{internal}
