% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_FREES_AGAINST_WT}
\alias{CREDIT_FREES_AGAINST_WT}
\title{Frees against weight (spoil component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_FREES_AGAINST_WT
}
\description{
Frees against weight (spoil component)
}
\keyword{internal}
