% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_GROUND_BALL_GETS_WT}
\alias{CREDIT_GROUND_BALL_GETS_WT}
\title{Ground ball gets weight (recv component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_GROUND_BALL_GETS_WT
}
\description{
Ground ball gets weight (recv component)
}
\keyword{internal}
