% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_HANDBALLS_WT}
\alias{CREDIT_HANDBALLS_WT}
\title{Handballs weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_HANDBALLS_WT
}
\description{
Handballs weight (disp component)
}
\keyword{internal}
