% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_HITOUT_WT}
\alias{CREDIT_HITOUT_WT}
\title{Hitout weight per hitout}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_HITOUT_WT
}
\description{
Hitout weight per hitout
}
\keyword{internal}
