% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_INTERCEPTS_WT}
\alias{CREDIT_INTERCEPTS_WT}
\title{Intercepts weight (spoil component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_INTERCEPTS_WT
}
\description{
Intercepts weight (spoil component)
}
\keyword{internal}
