% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_KICKS_WT}
\alias{CREDIT_KICKS_WT}
\title{Kicks weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_KICKS_WT
}
\description{
Kicks weight (disp component)
}
\keyword{internal}
