% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_POS_ADJ_QUANTILE_DISP}
\alias{CREDIT_POS_ADJ_QUANTILE_DISP}
\title{Position adjustment quantile for disposal}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_POS_ADJ_QUANTILE_DISP
}
\description{
Position adjustment quantile for disposal
}
\keyword{internal}
