% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_POS_ADJ_QUANTILE_HITOUT}
\alias{CREDIT_POS_ADJ_QUANTILE_HITOUT}
\title{Position adjustment quantile for hitout}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_POS_ADJ_QUANTILE_HITOUT
}
\description{
Position adjustment quantile for hitout
}
\keyword{internal}
