% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_POS_ADJ_QUANTILE_RECV}
\alias{CREDIT_POS_ADJ_QUANTILE_RECV}
\title{Position adjustment quantile for reception}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_POS_ADJ_QUANTILE_RECV
}
\description{
Position adjustment quantile for reception
}
\keyword{internal}
