% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_POS_ADJ_QUANTILE_SPOIL}
\alias{CREDIT_POS_ADJ_QUANTILE_SPOIL}
\title{Position adjustment quantile for spoil/tackle}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_POS_ADJ_QUANTILE_SPOIL
}
\description{
Position adjustment quantile for spoil/tackle
}
\keyword{internal}
