% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_PRESSURE_WT}
\alias{CREDIT_PRESSURE_WT}
\title{Pressure act weight}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_PRESSURE_WT
}
\description{
Pressure act weight
}
\keyword{internal}
