% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_REBOUND50S_WT}
\alias{CREDIT_REBOUND50S_WT}
\title{Rebound 50s weight (spoil component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_REBOUND50S_WT
}
\description{
Rebound 50s weight (spoil component)
}
\keyword{internal}
