% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_RECV_INTERCEPT_MARK_SCALE}
\alias{CREDIT_RECV_INTERCEPT_MARK_SCALE}
\title{Reception scaling factor for intercept marks (pos_team == -1 AND mark in PBP)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_RECV_INTERCEPT_MARK_SCALE
}
\description{
Reception scaling factor for intercept marks (pos_team == -1 AND mark in PBP)
}
\keyword{internal}
