% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_RECV_SCALE}
\alias{CREDIT_RECV_SCALE}
\title{Reception scaling factor}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_RECV_SCALE
}
\description{
Reception scaling factor
}
\keyword{internal}
