% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_RUCK_CONTEST_WT}
\alias{CREDIT_RUCK_CONTEST_WT}
\title{Ruck contest weight (subtracted)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_RUCK_CONTEST_WT
}
\description{
Ruck contest weight (subtracted)
}
\keyword{internal}
