% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_SCORE_INVOLVEMENTS_WT}
\alias{CREDIT_SCORE_INVOLVEMENTS_WT}
\title{Score involvements weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_SCORE_INVOLVEMENTS_WT
}
\description{
Score involvements weight (disp component)
}
\keyword{internal}
