% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_SHOTS_AT_GOAL_WT}
\alias{CREDIT_SHOTS_AT_GOAL_WT}
\title{Shots at goal weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_SHOTS_AT_GOAL_WT
}
\description{
Shots at goal weight (disp component)
}
\keyword{internal}
