% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_TACKLE_WT}
\alias{CREDIT_TACKLE_WT}
\title{Tackle weight per tackle}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_TACKLE_WT
}
\description{
Tackle weight per tackle
}
\keyword{internal}
