% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decimal_hour}
\alias{decimal_hour}
\title{Get mode of a vector}
\usage{
decimal_hour(datetime)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The mode (most frequent value) of the input vector.
}
\description{
This function returns the mode (most frequent value) of a vector.
}
