% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{determine_phase_of_play}
\alias{determine_phase_of_play}
\title{Determine Phase of Play}
\usage{
determine_phase_of_play(lag_desc_tot, description, throw_in)
}
\arguments{
\item{lag_desc_tot}{The lagged description of the play.}

\item{description}{The current description of the play.}

\item{throw_in}{Whether it's a throw-in.}
}
\value{
A factor representing the phase of play.
}
\description{
Determine Phase of Play
}
\keyword{internal}
