% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{determine_play_type}
\alias{determine_play_type}
\title{Determine Play Type}
\usage{
determine_play_type(description)
}
\arguments{
\item{description}{The description of the play.}
}
\value{
A factor representing the play type.
}
\description{
Determine Play Type
}
\keyword{internal}
