% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{determine_team_id_mdl}
\alias{determine_team_id_mdl}
\title{Determine Team ID Model}
\usage{
determine_team_id_mdl(throw_in, team_id)
}
\arguments{
\item{throw_in}{A vector indicating if the play is a throw-in.}

\item{team_id}{A vector of team IDs.}
}
\value{
A vector of determined team IDs.
}
\description{
Determine Team ID Model
}
\keyword{internal}
