% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DISK_CACHE_DEFAULT_AGE_DAYS}
\alias{DISK_CACHE_DEFAULT_AGE_DAYS}
\title{Default disk cache max age in days}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DISK_CACHE_DEFAULT_AGE_DAYS
}
\description{
Default disk cache max age in days
}
\keyword{internal}
