% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_skills.R
\name{.col_pctiles}
\alias{.col_pctiles}
\title{Compute percentile rank of a player's values within a reference data.table}
\usage{
.col_pctiles(dt, cols, player_row)
}
\description{
Compute percentile rank of a player's values within a reference data.table
}
\keyword{internal}
