% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_skills_data.R
\name{.compute_stat_rating_denominator}
\alias{.compute_stat_rating_denominator}
\title{Compute denominator vector from a column spec string}
\usage{
.compute_stat_rating_denominator(dt, denom_spec)
}
\arguments{
\item{dt}{A data.table.}

\item{denom_spec}{A string like "col" or "col1+col2".}
}
\value{
Numeric vector of denominators.
}
\description{
Supports "col1+col2" for summing multiple columns.
}
\keyword{internal}
