% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_skills.R
\name{.extract_player_cols}
\alias{.extract_player_cols}
\title{Extract column values from a single-row data.table, defaulting to NA}
\usage{
.extract_player_cols(player_row, col_names)
}
\description{
Extract column values from a single-row data.table, defaulting to NA
}
\keyword{internal}
