% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_api.R
\name{.fetch_all_player_details}
\alias{.fetch_all_player_details}
\title{Fetch player details for multiple seasons in one parallel batch}
\usage{
.fetch_all_player_details(seasons)
}
\arguments{
\item{seasons}{Numeric vector of years}
}
\value{
A tibble of player details across all seasons
}
\description{
Fetch player details for multiple seasons in one parallel batch
}
\keyword{internal}
