% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data_prep.R
\name{.fetch_forecast_weather}
\alias{.fetch_forecast_weather}
\title{Fetch forecast weather from Open-Meteo API}
\usage{
.fetch_forecast_weather(fixtures_upcoming, all_grounds)
}
\description{
Fetch forecast weather from Open-Meteo API
}
\keyword{internal}
