% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_match.R
\name{.fetch_match_lineups}
\alias{.fetch_match_lineups}
\title{Fetch team lineups for specific match IDs}
\usage{
.fetch_match_lineups(match_ids, season)
}
\arguments{
\item{match_ids}{Character vector of match IDs.}

\item{season}{Numeric season year.}
}
\value{
A data.frame of team lineup data with normalised column names.
}
\description{
Fetch team lineups for specific match IDs
}
\keyword{internal}
