% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_match.R
\name{.load_psr_coefs}
\alias{.load_psr_coefs}
\title{Load PSR coefficient CSV}
\usage{
.load_psr_coefs(filename)
}
\arguments{
\item{filename}{The coefficient CSV filename.}
}
\value{
A data.frame with stat_name, beta, sd columns, or NULL if not found.
}
\description{
Load PSR coefficient CSV
}
\keyword{internal}
