% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_api.R
\name{.parse_match_stats}
\alias{.parse_match_stats}
\title{Parse match player stats response}
\usage{
.parse_match_stats(resp, match_id)
}
\arguments{
\item{resp}{Parsed JSON from the playerStats endpoint}

\item{match_id}{The match providerId}
}
\value{
A tibble with player stat data
}
\description{
Parse match player stats response
}
\keyword{internal}
