% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_model.R
\name{.parse_start_time}
\alias{.parse_start_time}
\title{Parse start_time strings to POSIXct}
\usage{
.parse_start_time(x)
}
\arguments{
\item{x}{Character vector of start time strings}
}
\value{
POSIXct vector
}
\description{
Handles the local-time formatted strings stored in predictions data.
}
\keyword{internal}
