% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_match.R
\name{.recompute_cumulative}
\alias{.recompute_cumulative}
\title{Recompute cumulative home/away scores from interleaved events}
\usage{
.recompute_cumulative(team, score_value)
}
\description{
Recompute cumulative home/away scores from interleaved events
}
\keyword{internal}
