% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_match.R
\name{.resolve_match_params}
\alias{.resolve_match_params}
\title{Resolve estimate and predicted total from ratings or predictions}
\usage{
.resolve_match_params(
  home_team,
  away_team,
  season,
  team_ratings,
  predictions,
  home_advantage
)
}
\description{
Resolve estimate and predicted total from ratings or predictions
}
\keyword{internal}
