% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.should_log}
\alias{.should_log}
\title{Check if a message at the given severity should be emitted}
\usage{
.should_log(msg_level)
}
\description{
Check if a message at the given severity should be emitted
}
\keyword{internal}
