% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_train.R
\name{.strip_gam_models}
\alias{.strip_gam_models}
\title{Strip a list of GAM models for compact serialisation}
\usage{
.strip_gam_models(models)
}
\arguments{
\item{models}{Named list of gam/bam objects}
}
\value{
Same list with stripped models
}
\description{
Strip a list of GAM models for compact serialisation
}
\keyword{internal}
