% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg.R
\name{.summarise_match_xg}
\alias{.summarise_match_xg}
\title{Summarise shot-level PBP into match-level xG stats}
\usage{
.summarise_match_xg(pbp, quarter = 1:4)
}
\arguments{
\item{pbp}{Play-by-play data with xscore and shot_row columns}

\item{quarter}{Numeric vector of quarters to include}
}
\value{
Tibble with one row per match
}
\description{
Summarise shot-level PBP into match-level xG stats
}
\keyword{internal}
