% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rapm.R
\name{.team_rapm_modal_bucket}
\alias{.team_rapm_modal_bucket}
\title{Modal (most frequent) non-NA value -- used to pick one position_bucket for
a player whose bucket drifted match-to-match.}
\usage{
.team_rapm_modal_bucket(x)
}
\description{
Modal (most frequent) non-NA value -- used to pick one position_bucket for
a player whose bucket drifted match-to-match.
}
\keyword{internal}
