% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_api.R
\name{.validate_afl_comp}
\alias{.validate_afl_comp}
\title{Validate a comp code against the known set}
\usage{
.validate_afl_comp(comp)
}
\arguments{
\item{comp}{Character, e.g. "AFLM" or "AFLW"}
}
\description{
Validate a comp code against the known set
}
\keyword{internal}
