% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{.write_season_disk_cache}
\alias{.write_season_disk_cache}
\title{Write per-season disk cache for CFS data}
\usage{
.write_season_disk_cache(prefix, season, data)
}
\arguments{
\item{prefix}{Cache prefix}

\item{season}{Season year}

\item{data}{Data frame to cache}
}
\description{
Write per-season disk cache for CFS data
}
\keyword{internal}
