% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{ELO_HGA}
\alias{ELO_HGA}
\title{Elo home-ground advantage, in Elo points}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELO_HGA
}
\description{
Elo home-ground advantage, in Elo points
}
\keyword{internal}
