% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{ELO_K}
\alias{ELO_K}
\title{Elo update rate (K-factor)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ELO_K
}
\description{
Elo update rate (K-factor)
}
\keyword{internal}
