% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contests.R
\name{empty_contests_dt}
\alias{empty_contests_dt}
\title{Create an empty contests data.table with correct schema}
\usage{
empty_contests_dt()
}
\value{
An empty data.table
}
\description{
Create an empty contests data.table with correct schema
}
\keyword{internal}
