% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ep_model}
\alias{ep_model}
\title{Expected points model for game situations}
\format{
A data frame with X rows and Y variables:
\describe{
\item{variable1}{Description of variable1}
\item{variable2}{Description of variable2}
...
}
}
\source{
\url{http://example.com}
}
\usage{
ep_model
}
\description{
Expected points model for game situations
}
\keyword{datasets}
