% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPR_DECAY_CONTEST}
\alias{EPR_DECAY_CONTEST}
\title{Decay factor (in days) for contest component weighting}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPR_DECAY_CONTEST
}
\description{
Decay factor (in days) for contest component weighting
}
\keyword{internal}
