% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPR_DECAY_DEFAULT_DAYS}
\alias{EPR_DECAY_DEFAULT_DAYS}
\title{Default decay factor (in days) — legacy alias for backwards compatibility}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPR_DECAY_DEFAULT_DAYS
}
\description{
Default decay factor (in days) — legacy alias for backwards compatibility
}
\keyword{internal}
