% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPR_DECAY_SPOIL}
\alias{EPR_DECAY_SPOIL}
\title{Decay factor (in days) for spoil component weighting}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPR_DECAY_SPOIL
}
\description{
Decay factor (in days) for spoil component weighting
}
\keyword{internal}
