% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPR_LOADING_DEFAULT}
\alias{EPR_LOADING_DEFAULT}
\title{Loading factor for TORP calculations}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPR_LOADING_DEFAULT
}
\description{
Loading factor for TORP calculations
}
\keyword{internal}
