% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPR_PRIOR_GAMES_CONTEST}
\alias{EPR_PRIOR_GAMES_CONTEST}
\title{Prior games constant for contest ratings}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPR_PRIOR_GAMES_CONTEST
}
\description{
Prior games constant for contest ratings
}
\keyword{internal}
