% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPR_PRIOR_GAMES_DISP}
\alias{EPR_PRIOR_GAMES_DISP}
\title{Prior games constant for disposal ratings. See \code{EPR_PRIOR_GAMES_RECV}.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPR_PRIOR_GAMES_DISP
}
\description{
Prior games constant for disposal ratings. See \code{EPR_PRIOR_GAMES_RECV}.
}
\keyword{internal}
