% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_BEHINDS_WT}
\alias{EPV_BEHINDS_WT}
\title{Behinds weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_BEHINDS_WT
}
\description{
Behinds weight (disp component)
}
\keyword{internal}
