% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_BOUNCE_WT}
\alias{EPV_BOUNCE_WT}
\title{Bounce weight per bounce (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_BOUNCE_WT
}
\description{
Bounce weight per bounce (disp component)
}
\keyword{internal}
