% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_CLEARANCES_WT}
\alias{EPV_CLEARANCES_WT}
\title{Clearances weight (hitout component, currently disabled pending re-optimization)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_CLEARANCES_WT
}
\description{
Clearances weight (hitout component, currently disabled pending re-optimization)
}
\keyword{internal}
