% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_CONTESTED_POSS_WT}
\alias{EPV_CONTESTED_POSS_WT}
\title{Contested possessions weight (recv component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_CONTESTED_POSS_WT
}
\description{
Contested possessions weight (recv component)
}
\keyword{internal}
