% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_difficulty.R
\docType{data}
\name{EPV_DIFFICULTY_DESCS}
\alias{EPV_DIFFICULTY_DESCS}
\title{Disposal descriptions the difficulty split covers}
\format{
An object of class \code{character} of length 3.
}
\usage{
EPV_DIFFICULTY_DESCS
}
\description{
Disposal descriptions the difficulty split covers
}
\keyword{internal}
