% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_DISP_NEG_OFFSET}
\alias{EPV_DISP_NEG_OFFSET}
\title{Disposal EPV offset when defending (pos_team == -1)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_DISP_NEG_OFFSET
}
\description{
Disposal EPV offset when defending (pos_team == -1)
}
\keyword{internal}
