% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_DISP_SCALE}
\alias{EPV_DISP_SCALE}
\title{Disposal scaling factor}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_DISP_SCALE
}
\description{
Disposal scaling factor
}
\keyword{internal}
