% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_FREES_FOR_WT}
\alias{EPV_FREES_FOR_WT}
\title{Frees for weight (recv component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_FREES_FOR_WT
}
\description{
Frees for weight (recv component)
}
\keyword{internal}
