% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_GOAL_ASSISTS_WT}
\alias{EPV_GOAL_ASSISTS_WT}
\title{Goal assists weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_GOAL_ASSISTS_WT
}
\description{
Goal assists weight (disp component)
}
\keyword{internal}
