% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_GOALS_WT}
\alias{EPV_GOALS_WT}
\title{Goals weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_GOALS_WT
}
\description{
Goals weight (disp component)
}
\keyword{internal}
