% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_INSIDE50S_WT}
\alias{EPV_INSIDE50S_WT}
\title{Inside 50s weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_INSIDE50S_WT
}
\description{
Inside 50s weight (disp component)
}
\keyword{internal}
