% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_MARKS_INSIDE50_WT}
\alias{EPV_MARKS_INSIDE50_WT}
\title{Marks inside 50 weight (recv component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_MARKS_INSIDE50_WT
}
\description{
Marks inside 50 weight (recv component)
}
\keyword{internal}
