% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_MARKS_WT}
\alias{EPV_MARKS_WT}
\title{Total marks weight (recv component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_MARKS_WT
}
\description{
Total marks weight (recv component)
}
\keyword{internal}
