% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_METRES_GAINED_WT}
\alias{EPV_METRES_GAINED_WT}
\title{Metres gained weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_METRES_GAINED_WT
}
\description{
Metres gained weight (disp component)
}
\keyword{internal}
