% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_ONE_PERCENTERS_WT}
\alias{EPV_ONE_PERCENTERS_WT}
\title{One percenters weight (spoil component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_ONE_PERCENTERS_WT
}
\description{
One percenters weight (spoil component)
}
\keyword{internal}
