% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{EPV_OPP_PRIOR_GAMES}
\alias{EPV_OPP_PRIOR_GAMES}
\title{Pseudo-games at league average for EPV opponent profile shrinkage.
Pulls early-season teams toward league average.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_OPP_PRIOR_GAMES
}
\description{
Pseudo-games at league average for EPV opponent profile shrinkage.
Pulls early-season teams toward league average.
}
\keyword{internal}
