% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_POS_ADJ_QUANTILE_DISP}
\alias{EPV_POS_ADJ_QUANTILE_DISP}
\title{Position-group quantile adjustment for disposal component}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_POS_ADJ_QUANTILE_DISP
}
\description{
Position-group quantile adjustment for disposal component
}
\keyword{internal}
