% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_REBOUND50S_WT}
\alias{EPV_REBOUND50S_WT}
\title{Rebound 50s weight (spoil component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_REBOUND50S_WT
}
\description{
Rebound 50s weight (spoil component)
}
\keyword{internal}
