% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_RECV_NEG_OFFSET}
\alias{EPV_RECV_NEG_OFFSET}
\title{Reception offset when defending}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_RECV_NEG_OFFSET
}
\description{
Reception offset when defending
}
\keyword{internal}
