% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_RECV_POS_MULT}
\alias{EPV_RECV_POS_MULT}
\title{Reception multiplier when possessing (pos_team == 1)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_RECV_POS_MULT
}
\description{
Reception multiplier when possessing (pos_team == 1)
}
\keyword{internal}
