% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_RECV_POS_OFFSET}
\alias{EPV_RECV_POS_OFFSET}
\title{Reception offset when possessing}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_RECV_POS_OFFSET
}
\description{
Reception offset when possessing
}
\keyword{internal}
