% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_data.R
\docType{data}
\name{EPV_RELEVANT_DESCRIPTIONS}
\alias{EPV_RELEVANT_DESCRIPTIONS}
\title{Descriptions relevant for EPV modeling (the 27-description whitelist)
Used in clean_model_data_epv_dt() and filter_relevant_descriptions()}
\format{
An object of class \code{character} of length 27.
}
\usage{
EPV_RELEVANT_DESCRIPTIONS
}
\description{
Descriptions relevant for EPV modeling (the 27-description whitelist)
Used in clean_model_data_epv_dt() and filter_relevant_descriptions()
}
\keyword{internal}
