% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_SPOIL_WT}
\alias{EPV_SPOIL_WT}
\title{Spoil weight per spoil}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_SPOIL_WT
}
\description{
Spoil weight per spoil
}
\keyword{internal}
