% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_TURNOVERS_WT}
\alias{EPV_TURNOVERS_WT}
\title{Turnovers weight (disp component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_TURNOVERS_WT
}
\description{
Turnovers weight (disp component)
}
\keyword{internal}
