% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_UNCONTESTED_POSS_WT}
\alias{EPV_UNCONTESTED_POSS_WT}
\title{Uncontested possessions weight (recv component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_UNCONTESTED_POSS_WT
}
\description{
Uncontested possessions weight (recv component)
}
\keyword{internal}
