% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_WEIGHT_DECAY_DAYS}
\alias{EPV_WEIGHT_DECAY_DAYS}
\title{Decay factor (in days) for PBP-level game recency weighting in add_epv_vars()}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_WEIGHT_DECAY_DAYS
}
\description{
Decay factor (in days) for PBP-level game recency weighting in add_epv_vars()
}
\keyword{internal}
