% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FIELD_ZONE_SCORING_30}
\alias{FIELD_ZONE_SCORING_30}
\title{Scoring zone boundary at 30m from goal}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FIELD_ZONE_SCORING_30
}
\description{
Scoring zone boundary at 30m from goal
}
\keyword{internal}
