% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FIELD_ZONE_SCORING_80}
\alias{FIELD_ZONE_SCORING_80}
\title{Scoring zone boundary at 80m from goal}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FIELD_ZONE_SCORING_80
}
\description{
Scoring zone boundary at 80m from goal
}
\keyword{internal}
