% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{filter_game_data}
\alias{filter_game_data}
\title{Filter game data}
\usage{
filter_game_data(df, season_val, round_num, matchid, team)
}
\arguments{
\item{df}{Input data frame}

\item{season_val}{Season value}

\item{round_num}{Round number}

\item{matchid}{Match ID (NULL for no filtering)}

\item{team}{Team name (NULL for no filtering)}
}
\value{
Filtered data frame
}
\description{
Filter game data
}
