% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{filter_relevant_descriptions}
\alias{filter_relevant_descriptions}
\title{Filter Relevant Descriptions}
\usage{
filter_relevant_descriptions(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A filtered dataframe.
}
\description{
This function filters the dataframe to include only relevant play descriptions.
}
\keyword{internal}
