% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{generate_fixture_cache_key}
\alias{generate_fixture_cache_key}
\title{Generate cache key for fixture data}
\usage{
generate_fixture_cache_key(seasons, all)
}
\arguments{
\item{seasons}{Seasons parameter}

\item{all}{All parameter}
}
\value{
Character cache key
}
\description{
Generate cache key for fixture data
}
\keyword{internal}
