% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{generate_urls}
\alias{generate_urls}
\title{Generate URLs for data download}
\usage{
generate_urls(data_type, file_prefix, seasons, rounds = NULL)
}
\arguments{
\item{data_type}{Type of data (e.g., "chain-data", "pbp-data")}

\item{file_prefix}{Prefix for the file name}

\item{seasons}{A numeric vector of seasons}

\item{rounds}{A numeric vector of rounds (optional)}
}
\value{
A character vector of URLs
}
\description{
Generate URLs for data download
}
\keyword{internal}
