% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{get_afl_data_schemas}
\alias{get_afl_data_schemas}
\title{AFL Data Schema Definitions}
\usage{
get_afl_data_schemas()
}
\value{
List of data schema definitions
}
\description{
Defines expected schemas for different data types in the AFL analytics pipeline
}
\keyword{internal}
