% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_afl_week}
\alias{get_afl_week}
\title{Get AFL Week}
\usage{
get_afl_week(type = "current")
}
\arguments{
\item{type}{a logical to determine whether to use date-based heuristics to
determine current week, default TRUE}
}
\description{
Note that the date heuristic will count a new week starting on Thursdays, while
the schedule-based method will count a new week after the last game of the previous
week.
}
