% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_cache.R
\name{get_disk_cache_dir}
\alias{get_disk_cache_dir}
\title{Get Disk Cache Directory}
\usage{
get_disk_cache_dir()
}
\value{
Character string of the cache directory path
}
\description{
Returns the directory used for persistent disk caching.
Creates the directory if it doesn't exist.
}
\keyword{internal}
