% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_cache.R
\name{get_disk_cache_path}
\alias{get_disk_cache_path}
\title{Get Disk Cache Path for URL}
\usage{
get_disk_cache_path(url)
}
\arguments{
\item{url}{Character URL}
}
\value{
Character path to cache file
}
\description{
Returns the full path to the cached file for a given URL.
}
\keyword{internal}
