% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_variables.R
\name{get_epv_preds}
\alias{get_epv_preds}
\title{Get Expected Points Value Predictions}
\usage{
get_epv_preds(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with EPV predictions.
}
\description{
This function generates EPV predictions for the input data.
}
\keyword{internal}
