% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_functions.R
\name{get_game_chains}
\alias{get_game_chains}
\title{Get Game Chains}
\usage{
get_game_chains(match_id)
}
\arguments{
\item{match_id}{The ID of the match.}
}
\value{
A dataframe containing chain data for the specified game.
}
\description{
Retrieves chain data for a single game.
}
