% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_many_game_chains}
\alias{get_many_game_chains}
\title{Get Many Game Chains}
\usage{
get_many_game_chains(games_vector)
}
\arguments{
\item{games_vector}{A vector of game IDs.}
}
\value{
A dataframe containing chain data for all specified games.
}
\description{
This function is intended for internal use and may be unexported in a future release.
Retrieves chain data for multiple games.
}
