% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_model_cache_info}
\alias{get_model_cache_info}
\title{Get Model Cache Info}
\usage{
get_model_cache_info()
}
\value{
Character vector of cached model names
}
\description{
Returns information about cached models.
}
