% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_proportion_through_day}
\alias{get_proportion_through_day}
\title{Get Proportion Through Day}
\usage{
get_proportion_through_day(datetime)
}
\arguments{
\item{datetime}{A POSIXct datetime object.}
}
\value{
A numeric value representing the proportion of the day elapsed (0-1).
}
\description{
Calculates what proportion of the day has elapsed for a given datetime.
}
