% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_proportion_through_year}
\alias{get_proportion_through_year}
\title{Get Proportion Through Year}
\usage{
get_proportion_through_year(datetime)
}
\arguments{
\item{datetime}{A POSIXct datetime object.}
}
\value{
A numeric value representing the proportion of the year elapsed (0-1).
}
\description{
Calculates what proportion of the year has elapsed for a given datetime.
}
