% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_round_games}
\alias{get_round_games}
\title{Get Round Games}
\usage{
get_round_games(season, round, concluded_only = TRUE)
}
\arguments{
\item{season}{The AFL season year (numeric).}

\item{round}{The round number (numeric).}
}
\value{
A dataframe containing game data for the specified round.
}
\description{
Retrieves game data for a specific round in a season.
}
\keyword{internal}
