% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{get_season_data}
\alias{get_season_data}
\title{Get season data}
\usage{
get_season_data(season_val, round_num)
}
\arguments{
\item{season_val}{Season value (can be a vector)}

\item{round_num}{Round number}
}
\value{
Season data frame
}
\description{
Get season data
}
