% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_variables.R
\name{get_shot_result_preds}
\alias{get_shot_result_preds}
\title{Get Shot Result Predictions}
\usage{
get_shot_result_preds(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with shot result predictions.
}
\description{
This function generates shot result predictions for the input data.
}
\keyword{internal}
