% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_functions.R
\name{get_single_chain}
\alias{get_single_chain}
\title{Get Single Chain}
\usage{
get_single_chain(chains_t2, chain_number)
}
\arguments{
\item{chains_t2}{The chain data for a game.}

\item{chain_number}{The number of the chain to process.}
}
\value{
A dataframe containing data for the specified chain.
}
\description{
Processes a single chain from the game data.
}
