% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{get_torp_data_repo}
\alias{get_torp_data_repo}
\title{Get TORP Data Repository}
\usage{
get_torp_data_repo()
}
\value{
Character string of the repository in format "owner/repo"
}
\description{
Returns the repository used for TORP data downloads. Can be configured
via the TORP_DATA_REPO environment variable or package options.
}
\keyword{internal}
