% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_functions.R
\name{get_week_chains}
\alias{get_week_chains}
\title{Get Week Chains}
\usage{
get_week_chains(season, roundnum)
}
\arguments{
\item{season}{The AFL season year (numeric).}

\item{roundnum}{The round number (numeric).}
}
\value{
A dataframe containing chain data for the specified week.
}
\description{
Retrieves chain data for a specific week (round) in a season.
}
