% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{INJURY_DISCOUNT_FLOOR}
\alias{INJURY_DISCOUNT_FLOOR}
\title{Floor for injury discount scaling
Discount drops 0.01 per week from INJURY_KNOWN_DISCOUNT, floored here.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
INJURY_DISCOUNT_FLOOR
}
\description{
Floor for injury discount scaling
Discount drops 0.01 per week from INJURY_KNOWN_DISCOUNT, floored here.
}
\keyword{internal}
