% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is_cache_valid}
\alias{is_cache_valid}
\title{Check if cache entry is valid}
\usage{
is_cache_valid(cache_entry, cache_ttl)
}
\arguments{
\item{cache_entry}{Cache entry object}

\item{cache_ttl}{Time-to-live in seconds}
}
\value{
Logical indicating if cache is valid
}
\description{
Check if cache entry is valid
}
\keyword{internal}
