% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{is_consecutive_throw_in_same_side}
\alias{is_consecutive_throw_in_same_side}
\title{Check if consecutive throw-ins should be mirrored based on team and position}
\usage{
is_consecutive_throw_in_same_side(throw_in, team_id_mdl, x)
}
\arguments{
\item{throw_in}{A vector indicating if the play is a throw-in.}

\item{team_id_mdl}{A vector of team IDs for modeling.}

\item{x}{A vector of x-coordinates.}
}
\value{
Logical vector indicating mirror condition
}
\description{
Check if consecutive throw-ins should be mirrored based on team and position
}
\keyword{internal}
