% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{is_consecutive_throw_in_side_change}
\alias{is_consecutive_throw_in_side_change}
\title{Check if consecutive throw-ins should be mirrored based on position change}
\usage{
is_consecutive_throw_in_side_change(throw_in, team_id_mdl, x)
}
\arguments{
\item{throw_in}{A vector indicating if the play is a throw-in.}

\item{team_id_mdl}{A vector of team IDs for modeling.}

\item{x}{A vector of x-coordinates.}
}
\value{
Logical vector indicating mirror condition
}
\description{
Check if consecutive throw-ins should be mirrored based on position change
}
\keyword{internal}
