% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{is_current_throw_in_team_change}
\alias{is_current_throw_in_team_change}
\title{Check if current throw-in should be mirrored based on team change}
\usage{
is_current_throw_in_team_change(throw_in, team_id_mdl)
}
\arguments{
\item{throw_in}{A vector indicating if the play is a throw-in.}

\item{team_id_mdl}{A vector of team IDs for modeling.}
}
\value{
Logical vector indicating mirror condition
}
\description{
Check if current throw-in should be mirrored based on team change
}
\keyword{internal}
