% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{is_previous_throw_in_affecting}
\alias{is_previous_throw_in_affecting}
\title{Check if previous throw-in affects current play mirroring}
\usage{
is_previous_throw_in_affecting(throw_in, team_id_mdl, x)
}
\arguments{
\item{throw_in}{A vector indicating if the play is a throw-in.}

\item{team_id_mdl}{A vector of team IDs for modeling.}

\item{x}{A vector of x-coordinates.}
}
\value{
Logical vector indicating mirror condition
}
\description{
Check if previous throw-in affects current play mirroring
}
\keyword{internal}
