% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{is_throw_in_affecting_future_position}
\alias{is_throw_in_affecting_future_position}
\title{Check if throw-in affects future position}
\usage{
is_throw_in_affecting_future_position(throw_in, team_id_mdl, x)
}
\arguments{
\item{throw_in}{A vector indicating if the play is a throw-in.}

\item{team_id_mdl}{A vector of team IDs for modeling.}

\item{x}{A vector of x-coordinates.}
}
\value{
Logical vector indicating mirror condition
}
\description{
Check if throw-in affects future position
}
\keyword{internal}
