% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{is_throw_in_two_ago_affecting}
\alias{is_throw_in_two_ago_affecting}
\title{Check if throw-in two plays ago affects current play}
\usage{
is_throw_in_two_ago_affecting(throw_in, team_id_mdl, x)
}
\arguments{
\item{throw_in}{A vector indicating if the play is a throw-in.}

\item{team_id_mdl}{A vector of team IDs for modeling.}

\item{x}{A vector of x-coordinates.}
}
\value{
Logical vector indicating mirror condition
}
\description{
Check if throw-in two plays ago affects current play
}
\keyword{internal}
