% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chains.R
\name{load_chains}
\alias{load_chains}
\title{Load Chains}
\usage{
load_chains(
  seasons = lubridate::year(Sys.Date()),
  weeks = 1:35,
  file_type = "rds"
)
}
\arguments{
\item{file_type}{csv or rds (soon to come)}
}
\description{
These functions allows you to load AFL chains quickly
}
\examples{
\dontrun{
load_chains()
}

}
\keyword{AFL}
