% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MATCH_COMBO_POS_MAP}
\alias{MATCH_COMBO_POS_MAP}
\title{Combined position map for match models (structural pairs/trios)}
\format{
An object of class \code{list} of length 8.
}
\usage{
MATCH_COMBO_POS_MAP
}
\description{
Combined position map for match models (structural pairs/trios)
}
\keyword{internal}
