% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MATCH_DEFAULT_REST_DAYS}
\alias{MATCH_DEFAULT_REST_DAYS}
\title{Default days rest when rest data is unavailable}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MATCH_DEFAULT_REST_DAYS
}
\description{
Default days rest when rest data is unavailable
}
\keyword{internal}
