% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{MATCH_INDIVIDUAL_POS}
\alias{MATCH_INDIVIDUAL_POS}
\title{Individual field position codes for match models (18 on-field positions)}
\format{
An object of class \code{character} of length 18.
}
\usage{
MATCH_INDIVIDUAL_POS
}
\description{
Individual field position codes for match models (18 on-field positions)
}
\keyword{internal}
