% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MATCH_LOG_DIST_DEFAULT}
\alias{MATCH_LOG_DIST_DEFAULT}
\title{Default log-distance value when venue coordinates are missing}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MATCH_LOG_DIST_DEFAULT
}
\description{
Default log-distance value when venue coordinates are missing
}
\keyword{internal}
