% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{MATCH_LOG_DIST_OFFSET}
\alias{MATCH_LOG_DIST_OFFSET}
\title{Offset added to Haversine distance (metres) before log-transform
Prevents log(0) for home games and dampens the curve for short trips}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MATCH_LOG_DIST_OFFSET
}
\description{
Offset added to Haversine distance (metres) before log-transform
Prevents log(0) for home games and dampens the curve for short trips
}
\keyword{internal}
