% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{MATCH_MIN_DATA_ROUND}
\alias{MATCH_MIN_DATA_ROUND}
\title{Earliest round in MATCH_MIN_DATA_SEASON with reliable data}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MATCH_MIN_DATA_ROUND
}
\description{
Earliest round in MATCH_MIN_DATA_SEASON with reliable data
}
\keyword{internal}
