% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{MATCH_MIN_DATA_SEASON}
\alias{MATCH_MIN_DATA_SEASON}
\title{Earliest season with reliable TORP + xG data for match modelling}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MATCH_MIN_DATA_SEASON
}
\description{
Earliest season with reliable TORP + xG data for match modelling
}
\keyword{internal}
