% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MATCH_PHASE_MAP}
\alias{MATCH_PHASE_MAP}
\title{Phase groupings for match model position columns
Maps field position (position.x) to broad phase categories}
\format{
An object of class \code{list} of length 4.
}
\usage{
MATCH_PHASE_MAP
}
\description{
Phase groupings for match model position columns
Maps field position (position.x) to broad phase categories
}
\keyword{internal}
