% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{MATCH_POS_COLS}
\alias{MATCH_POS_COLS}
\title{All generated position column names for match model aggregation}
\format{
An object of class \code{character} of length 43.
}
\usage{
MATCH_POS_COLS
}
\description{
All generated position column names for match model aggregation
}
\keyword{internal}
