% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{MATCH_POS_GROUP_MAP}
\alias{MATCH_POS_GROUP_MAP}
\title{Position group aggregation map for match models}
\format{
An object of class \code{list} of length 6.
}
\usage{
MATCH_POS_GROUP_MAP
}
\description{
Position group aggregation map for match models
}
\keyword{internal}
