% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{MATCH_SIM_AVG_CONV_RATE}
\alias{MATCH_SIM_AVG_CONV_RATE}
\title{Average goal conversion rate (goals / shots, empirical 2021-2025)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MATCH_SIM_AVG_CONV_RATE
}
\description{
Average goal conversion rate (goals / shots, empirical 2021-2025)
}
\keyword{internal}
