% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MATCH_SIM_AVG_SHOTS_PER_QTR}
\alias{MATCH_SIM_AVG_SHOTS_PER_QTR}
\title{Average shots per team per quarter (empirical from 2021-2025 PBP)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MATCH_SIM_AVG_SHOTS_PER_QTR
}
\description{
Average shots per team per quarter (empirical from 2021-2025 PBP)
}
\keyword{internal}
