% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MATCH_SIM_DEFAULT_N}
\alias{MATCH_SIM_DEFAULT_N}
\title{Default number of Monte Carlo simulations for match simulation}
\format{
An object of class \code{integer} of length 1.
}
\usage{
MATCH_SIM_DEFAULT_N
}
\description{
Default number of Monte Carlo simulations for match simulation
}
\keyword{internal}
