% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MATCH_WEIGHT_DECAY_DAYS}
\alias{MATCH_WEIGHT_DECAY_DAYS}
\title{Exponential decay half-life in days for match model time-weighting}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MATCH_WEIGHT_DECAY_DAYS
}
\description{
Exponential decay half-life in days for match model time-weighting
}
\keyword{internal}
