% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_xg_functions.R
\name{match_xgs}
\alias{match_xgs}
\title{add ep vars}
\usage{
match_xgs(season, round, match_id = NA)
}
\arguments{
\item{df}{}
}
\description{
add ep vars
}
\examples{
\dontrun{
add_epv_vars(df)
}
}
