% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{NP_BALL_WINNER_SHARE_BY_ACT_DEFAULT}
\alias{NP_BALL_WINNER_SHARE_BY_ACT_DEFAULT}
\title{Fallback for \code{NP_BALL_WINNER_SHARE_BY_ACT}}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
NP_BALL_WINNER_SHARE_BY_ACT_DEFAULT
}
\description{
Fallback for \code{NP_BALL_WINNER_SHARE_BY_ACT}
}
\keyword{internal}
