% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{NP_CONTEST_WINNER_SHARE_DEFAULT}
\alias{NP_CONTEST_WINNER_SHARE_DEFAULT}
\title{Fallback for \code{NP_CONTEST_WINNER_SHARE}}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
NP_CONTEST_WINNER_SHARE_DEFAULT
}
\description{
Fallback for \code{NP_CONTEST_WINNER_SHARE}
}
\keyword{internal}
