% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{NP_DISPOSAL_DESCS}
\alias{NP_DISPOSAL_DESCS}
\title{PBP descriptions that are disposals for net-points purposes}
\format{
An object of class \code{character} of length 2.
}
\usage{
NP_DISPOSAL_DESCS
}
\description{
PBP descriptions that are disposals for net-points purposes
}
\keyword{internal}
