% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{NP_STOPPAGE_BAND_M}
\alias{NP_STOPPAGE_BAND_M}
\title{Width, in metres, of the location bands the stoppage baseline is averaged in}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
NP_STOPPAGE_BAND_M
}
\description{
Width, in metres, of the location bands the stoppage baseline is averaged in
}
\keyword{internal}
