% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left operand}

\item{y}{Right operand}
}
\value{
x if not NULL, otherwise y
}
\description{
Returns the left operand if not NULL, otherwise the right operand
}
