% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{OPEN_METEO_RATE_LIMIT_SECONDS}
\alias{OPEN_METEO_RATE_LIMIT_SECONDS}
\title{Rate limit delay in seconds for Open-Meteo API calls}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
OPEN_METEO_RATE_LIMIT_SECONDS
}
\description{
Rate limit delay in seconds for Open-Meteo API calls
}
\keyword{internal}
