% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{OPP_ADJ_FACTOR_CAP}
\alias{OPP_ADJ_FACTOR_CAP}
\title{Floor and ceiling for opponent adjustment factor.
Prevents extreme adjustments from small samples.}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
OPP_ADJ_FACTOR_CAP
}
\description{
Floor and ceiling for opponent adjustment factor.
Prevents extreme adjustments from small samples.
}
\keyword{internal}
