% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{OPP_ADJ_LAMBDA_DECAY}
\alias{OPP_ADJ_LAMBDA_DECAY}
\title{Decay rate (per day) for opponent defensive profile recency weighting.
~231-day half-life, slightly faster than stat rating decay.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
OPP_ADJ_LAMBDA_DECAY
}
\description{
Decay rate (per day) for opponent defensive profile recency weighting.
~231-day half-life, slightly faster than stat rating decay.
}
\keyword{internal}
