% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_engines.R
\name{parquet_from_url}
\alias{parquet_from_url}
\title{Load parquet file from a remote connection}
\usage{
parquet_from_url(url)
}
\arguments{
\item{url}{A character URL}
}
\value{
A data frame
}
\description{
This function is intended for internal use and may be unexported in a future release.
}
