% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_game_ratings.R
\name{player_season_ratings}
\alias{player_season_ratings}
\title{Get season total ratings}
\usage{
player_season_ratings(season_val = get_afl_season(), round_num = NULL)
}
\arguments{
\item{season_val}{The season to calculate ratings for. Default is the current season.}

\item{round_num}{Deprecated and ignored. Retained for backwards compatibility.}
}
\value{
A data frame containing player season total ratings.
}
\description{
Convenience wrapper around \code{\link[=load_player_season_ratings]{load_player_season_ratings()}} with filtering.
}
