% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plyr_tm_df}
\alias{plyr_tm_df}
\title{Player team database}
\format{
A data frame with X rows and Y variables:
\describe{
\item{player_id}{Unique identifier for each player}
\item{team_id}{Unique identifier for each team}
\item{season}{Season of play}
...
}
}
\source{
\url{http://example.com}
}
\usage{
plyr_tm_df
}
\description{
Player team database
}
\keyword{datasets}
