% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{POSITION_AVG_TOG_DEFAULT}
\alias{POSITION_AVG_TOG_DEFAULT}
\title{Default TOG fraction for unknown field positions}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
POSITION_AVG_TOG_DEFAULT
}
\description{
Default TOG fraction for unknown field positions
}
\keyword{internal}
