% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_features.R
\name{predict_shot_probabilities}
\alias{predict_shot_probabilities}
\title{Predict Shot Probabilities}
\usage{
predict_shot_probabilities(model, new_data)
}
\arguments{
\item{model}{A fitted shot model.}

\item{new_data}{A dataframe or model matrix containing new shot data.}
}
\value{
A vector of predicted probabilities.
}
\description{
Predict Shot Probabilities
}
\keyword{internal}
