% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_spm.R
\name{predict_team_spm}
\alias{predict_team_spm}
\title{Apply a fitted SPM to ANY player with box-score data -- including players
RAPM pooled into a replacement-level column for lack of games/minutes.}
\usage{
predict_team_spm(spm_fit, spm_features)
}
\arguments{
\item{spm_fit}{Output of \code{fit_team_spm()}.}

\item{spm_features}{Output of \code{build_team_spm_features()}.}
}
\value{
data.table(player_id, n_games, total_tog_minutes, spm_offense,
spm_defense, spm_net).
}
\description{
Apply a fitted SPM to ANY player with box-score data -- including players
RAPM pooled into a replacement-level column for lack of games/minutes.
}
\keyword{internal}
