% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_models.R
\name{predict_wp_ensemble_baseline}
\alias{predict_wp_ensemble_baseline}
\title{Ensemble Baseline Win Probability Prediction}
\usage{
predict_wp_ensemble_baseline(data)
}
\arguments{
\item{data}{Input data}
}
\value{
Vector of ensemble win probability predictions
}
\description{
Combines multiple baseline models using simple averaging
}
