% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_models.R
\name{predict_wp_expected_points}
\alias{predict_wp_expected_points}
\title{Expected Points Baseline Model}
\usage{
predict_wp_expected_points(train_data, pred_data)
}
\arguments{
\item{train_data}{Training data}

\item{pred_data}{Data to make predictions on}
}
\value{
Vector of win probability predictions
}
\description{
Uses expected points differential as the primary predictor
}
