% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_models.R
\name{predict_wp_gam_baseline}
\alias{predict_wp_gam_baseline}
\title{GAM Baseline Model}
\usage{
predict_wp_gam_baseline(data, pred_data = NULL)
}
\arguments{
\item{data}{Data to make predictions on (if only one parameter provided)}

\item{pred_data}{Optional - data to make predictions on}
}
\value{
Vector of win probability predictions
}
\description{
Generalized Additive Model baseline with smooth terms
}
