% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_models.R
\name{predict_wp_time_score}
\alias{predict_wp_time_score}
\title{Time-Aware Baseline Model for Win Probability}
\usage{
predict_wp_time_score(train_data, pred_data)
}
\arguments{
\item{train_data}{Training data}

\item{pred_data}{Data to make predictions on}
}
\value{
Vector of win probability predictions
}
\description{
Logistic regression using score difference and time remaining
}
