% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_models.R
\name{prepare_calibration_plot}
\alias{prepare_calibration_plot}
\title{Generate Calibration Plot Data}
\usage{
prepare_calibration_plot(calibration_results)
}
\arguments{
\item{calibration_results}{Results from assess_model_calibration}
}
\value{
ggplot2-ready dataframe for calibration plots
}
\description{
Prepares data for creating reliability/calibration plots
}
