% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{prepare_shot_model_data}
\alias{prepare_shot_model_data}
\title{Prepare Shot Data for Modeling}
\usage{
prepare_shot_model_data(df)
}
\arguments{
\item{df}{A dataframe containing raw shot data.}
}
\value{
A dataframe ready for shot modeling.
}
\description{
Prepare Shot Data for Modeling
}
