% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contests.R
\name{print.torp_head_to_head}
\alias{print.torp_head_to_head}
\title{Print Head-to-Head Results}
\usage{
\method{print}{torp_head_to_head}(x, ...)
}
\arguments{
\item{x}{A \code{torp_head_to_head} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Head-to-Head Results
}
