% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_profile.R
\name{print.torp_player_profile}
\alias{print.torp_player_profile}
\title{Print a player profile}
\usage{
\method{print}{torp_player_profile}(x, ...)
}
\arguments{
\item{x}{A \code{torp_player_profile} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a player profile
}
