% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder.R
\name{print.torp_sim_results}
\alias{print.torp_sim_results}
\title{Print simulation results}
\usage{
\method{print}{torp_sim_results}(x, ...)
}
\arguments{
\item{x}{A \code{torp_sim_results} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print simulation results
}
