% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_skills.R
\name{print.torp_skill_profile}
\alias{print.torp_skill_profile}
\title{Print a player skill profile}
\usage{
\method{print}{torp_skill_profile}(x, ...)
}
\arguments{
\item{x}{A \code{torp_skill_profile} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a player skill profile
}
