% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progressively}
\alias{progressively}
\title{Progressively apply a function}
\usage{
progressively(f, p = NULL)
}
\arguments{
\item{f}{A function to add progressor functionality to.}

\item{p}{A function such as one created by \code{progressr::progressor()} - also accepts purrr-style lambda functions.}
}
\value{
A function that does the same as \code{f} but calls \code{p()} after each iteration.
}
\description{
This function helps add progress-reporting to any function. Given function \code{f()} and progressor \code{p()},
it will return a new function that calls \code{f()} and then calls \code{p()} after every iteration.
}
