% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PSR_PRIOR_RATE}
\alias{PSR_PRIOR_RATE}
\title{PSR prior rate for replacement-level players}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PSR_PRIOR_RATE
}
\description{
Players without enough skill history to compute PSR are assigned this
replacement-level value. Based on empirical mean PSR of players with
1-10 career games (~-1.5 to -2.0).
}
\keyword{internal}
