% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RATING_DECAY_DEFAULT_DAYS}
\alias{RATING_DECAY_DEFAULT_DAYS}
\title{Default decay factor (in days) for weighting historical games}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_DECAY_DEFAULT_DAYS
}
\description{
Default decay factor (in days) for weighting historical games
}
\keyword{internal}
