% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RATING_DECAY_DISP}
\alias{RATING_DECAY_DISP}
\title{Decay factor (in days) for disposal component weighting}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_DECAY_DISP
}
\description{
Decay factor (in days) for disposal component weighting
}
\keyword{internal}
