% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RATING_PRIOR_GAMES_DISP}
\alias{RATING_PRIOR_GAMES_DISP}
\title{Prior games constant for disposal ratings}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_PRIOR_GAMES_DISP
}
\description{
Prior games constant for disposal ratings
}
\keyword{internal}
