% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RATING_PRIOR_GAMES_HITOUT}
\alias{RATING_PRIOR_GAMES_HITOUT}
\title{Prior games constant for hitout ratings}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_PRIOR_GAMES_HITOUT
}
\description{
Prior games constant for hitout ratings
}
\keyword{internal}
