% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RATING_PRIOR_GAMES_RECV}
\alias{RATING_PRIOR_GAMES_RECV}
\title{Prior games constant for receiving ratings}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_PRIOR_GAMES_RECV
}
\description{
Prior games constant for receiving ratings
}
\keyword{internal}
