% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RATING_PRIOR_GAMES_SPOIL}
\alias{RATING_PRIOR_GAMES_SPOIL}
\title{Prior games constant for spoil ratings}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_PRIOR_GAMES_SPOIL
}
\description{
Prior games constant for spoil ratings
}
\keyword{internal}
