% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RATING_PRIOR_RATE_DISP}
\alias{RATING_PRIOR_RATE_DISP}
\title{Prior rate for disposal component (shrinkage target per weighted game)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_PRIOR_RATE_DISP
}
\description{
Prior rate for disposal component (shrinkage target per weighted game)
}
\keyword{internal}
