% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RATING_SPOIL_MULTIPLIER}
\alias{RATING_SPOIL_MULTIPLIER}
\title{Spoil rating multiplier}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RATING_SPOIL_MULTIPLIER
}
\description{
Spoil rating multiplier
}
\keyword{internal}
