% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{rds_from_url}
\alias{rds_from_url}
\title{Load .rds file from a remote connection}
\usage{
rds_from_url(url)
}
\arguments{
\item{url}{A character URL}
}
\value{
A data frame as created by \code{\link[=readRDS]{readRDS()}}
}
\description{
Load .rds file from a remote connection
}
