% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{read_local_parquet}
\alias{read_local_parquet}
\title{Read a Parquet File from Local Storage}
\usage{
read_local_parquet(url)
}
\arguments{
\item{url}{Character URL used as the key to find the local file}
}
\value{
Data frame if local file exists, NULL otherwise
}
\description{
Read a Parquet File from Local Storage
}
\keyword{internal}
