% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{RELEASE_ASSET_CACHE_TTL_SECS}
\alias{RELEASE_ASSET_CACHE_TTL_SECS}
\title{Default release-asset listing cache TTL in seconds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RELEASE_ASSET_CACHE_TTL_SECS
}
\description{
Default release-asset listing cache TTL in seconds
}
\keyword{internal}
