% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{ROLE_BENCH_SLOTS}
\alias{ROLE_BENCH_SLOTS}
\title{Starting slots that are not roles}
\format{
An object of class \code{character} of length 3.
}
\usage{
ROLE_BENCH_SLOTS
}
\description{
Starting slots that are not roles
}
\keyword{internal}
