% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{select_afl_model_vars}
\alias{select_afl_model_vars}
\title{Select AFL Match Prediction Model Variables}
\usage{
select_afl_model_vars(df)
}
\arguments{
\item{df}{A dataframe containing AFL team model data}
}
\value{
A dataframe with selected variables for AFL modeling
}
\description{
Select AFL Match Prediction Model Variables
}
