% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{select_shot_model_vars}
\alias{select_shot_model_vars}
\title{Select Shot Model Variables}
\usage{
select_shot_model_vars(df)
}
\arguments{
\item{df}{A dataframe containing cleaned shot data.}
}
\value{
A dataframe with selected variables for shot modeling.
}
\description{
Select Shot Model Variables
}
