% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{select_wp_model_vars}
\alias{select_wp_model_vars}
\title{Select WP Model Variables}
\usage{
select_wp_model_vars(df)
}
\arguments{
\item{df}{A dataframe containing cleaned play-by-play data.}
}
\value{
A dataframe with selected variables for WP modeling.
}
\description{
Select WP Model Variables
}
