% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{set_torp_data_repo}
\alias{set_torp_data_repo}
\title{Set TORP Data Repository}
\usage{
set_torp_data_repo(repo)
}
\arguments{
\item{repo}{Character string of the repository in format "owner/repo"}
}
\description{
Sets the repository used for TORP data downloads in the current session.
}
