% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shot_player_df}
\alias{shot_player_df}
\title{Player shot data}
\format{
A data frame with X rows and Y variables:
\describe{
\item{player_id}{Unique identifier for each player}
\item{shot_id}{Unique identifier for each shot}
\item{x_coordinate}{X-coordinate of the shot}
\item{y_coordinate}{Y-coordinate of the shot}
\item{outcome}{Outcome of the shot}
...
}
}
\source{
\url{http://example.com}
}
\usage{
shot_player_df
}
\description{
Player shot data
}
\keyword{datasets}
