% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{SIM_AVG_TOTAL}
\alias{SIM_AVG_TOTAL}
\title{Fallback mean combined score per game}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_AVG_TOTAL
}
\description{
Fallback mean combined score per game
}
\keyword{internal}
