% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{SIM_DEFAULT_N}
\alias{SIM_DEFAULT_N}
\title{Default number of simulations}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_DEFAULT_N
}
\description{
Default number of simulations
}
\keyword{internal}
