% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIM_GF_HOME_ADVANTAGE}
\alias{SIM_GF_HOME_ADVANTAGE}
\title{Grand Final home advantage (MCG = roughly neutral)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_GF_HOME_ADVANTAGE
}
\description{
Grand Final home advantage (MCG = roughly neutral)
}
\keyword{internal}
