% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIM_GF_VENUE}
\alias{SIM_GF_VENUE}
\title{Grand Final venue (standardised name via torp_replace_venues)}
\format{
An object of class \code{character} of length 1.
}
\usage{
SIM_GF_VENUE
}
\description{
Grand Final venue (standardised name via torp_replace_venues)
}
\keyword{internal}
