% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIM_HOME_ADVANTAGE}
\alias{SIM_HOME_ADVANTAGE}
\title{Home ground advantage in points}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_HOME_ADVANTAGE
}
\description{
Home ground advantage in points
}
\keyword{internal}
