% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIM_INJURY_DISCOUNT}
\alias{SIM_INJURY_DISCOUNT}
\title{Injury discount applied to team rating sums
Accounts for typical player availability (injuries, rest, rotation).}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_INJURY_DISCOUNT
}
\description{
Injury discount applied to team rating sums
Accounts for typical player availability (injuries, rest, rotation).
}
\keyword{internal}
