% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{SIM_INJURY_SEASON_LATE}
\alias{SIM_INJURY_SEASON_LATE}
\title{"Late season" estimated return maps to round 18}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SIM_INJURY_SEASON_LATE
}
\description{
"Late season" estimated return maps to round 18
}
\keyword{internal}
