% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{SIM_INJURY_SEASON_MID}
\alias{SIM_INJURY_SEASON_MID}
\title{"Mid-season" estimated return maps to round 12}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SIM_INJURY_SEASON_MID
}
\description{
"Mid-season" estimated return maps to round 12
}
\keyword{internal}
