% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIM_INJURY_SECOND_HALF}
\alias{SIM_INJURY_SECOND_HALF}
\title{"Second half of YYYY" estimated return maps to round 13}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SIM_INJURY_SECOND_HALF
}
\description{
"Second half of YYYY" estimated return maps to round 13
}
\keyword{internal}
