% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{SIM_INJURY_TBC_BUFFER}
\alias{SIM_INJURY_TBC_BUFFER}
\title{TBC/unknown injuries: assume back in ~3 rounds}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SIM_INJURY_TBC_BUFFER
}
\description{
TBC/unknown injuries: assume back in ~3 rounds
}
\keyword{internal}
