% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIM_NOISE_SD}
\alias{SIM_NOISE_SD}
\title{Noise standard deviation for match simulation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_NOISE_SD
}
\description{
Noise standard deviation for match simulation
}
\keyword{internal}
