% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helpers.R
\name{sim_season}
\alias{sim_season}
\title{Simulate a season of games}
\usage{
sim_season(sim_teams, sim_games)
}
\arguments{
\item{sim_teams}{A data frame containing team ratings.}

\item{sim_games}{A data frame containing fixture data.}
}
\value{
A data frame of simulated game results.
}
\description{
This function simulates a season of games based on team ratings and fixture data.
}
