% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIM_TOP_N_PLAYERS}
\alias{SIM_TOP_N_PLAYERS}
\title{Top N players per team for rating aggregation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_TOP_N_PLAYERS
}
\description{
Top N players per team for rating aggregation
}
\keyword{internal}
