% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sim.R
\docType{data}
\name{SIM_TOTAL_SD}
\alias{SIM_TOTAL_SD}
\title{SD of combined score noise}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_TOTAL_SD
}
\description{
SD of combined score noise
}
\keyword{internal}
