% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SIM_WP_SCALING_FACTOR}
\alias{SIM_WP_SCALING_FACTOR}
\title{Win probability scaling factor for simulations}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SIM_WP_SCALING_FACTOR
}
\description{
Win probability scaling factor for simulations
}
\keyword{internal}
