% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SKILL_LAMBDA_EFFICIENCY_DEFAULT}
\alias{SKILL_LAMBDA_EFFICIENCY_DEFAULT}
\title{Default exponential decay rate for efficiency stats (per day)
Half-life = ln(2) / 0.0013 ~ 533 days}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SKILL_LAMBDA_EFFICIENCY_DEFAULT
}
\description{
Default exponential decay rate for efficiency stats (per day)
Half-life = ln(2) / 0.0013 ~ 533 days
}
\keyword{internal}
