% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SKILL_LAMBDA_RATE_DEFAULT}
\alias{SKILL_LAMBDA_RATE_DEFAULT}
\title{Default exponential decay rate for rate stats (per day)
Half-life = ln(2) / 0.0019 ~ 365 days}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SKILL_LAMBDA_RATE_DEFAULT
}
\description{
Default exponential decay rate for rate stats (per day)
Half-life = ln(2) / 0.0019 ~ 365 days
}
\keyword{internal}
