% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SKILL_MIN_GAMES}
\alias{SKILL_MIN_GAMES}
\title{Minimum weighted games for a player to appear in skill output}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SKILL_MIN_GAMES
}
\description{
Minimum weighted games for a player to appear in skill output
}
\keyword{internal}
