% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{STAT_RATING_CREDIBLE_LEVEL}
\alias{STAT_RATING_CREDIBLE_LEVEL}
\alias{SKILL_CREDIBLE_LEVEL}
\title{Credible interval width (0.80 = 80\% CI)}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
STAT_RATING_CREDIBLE_LEVEL

SKILL_CREDIBLE_LEVEL
}
\description{
Credible interval width (0.80 = 80\% CI)
}
\keyword{internal}
