% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STAT_RATING_MIN_GAMES}
\alias{STAT_RATING_MIN_GAMES}
\alias{SKILL_MIN_GAMES}
\title{Minimum weighted games for a player to appear in stat rating output}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
STAT_RATING_MIN_GAMES

SKILL_MIN_GAMES
}
\description{
Minimum weighted games for a player to appear in stat rating output
}
\keyword{internal}
