% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STAT_RATING_PRIOR_ATTEMPTS_DEFAULT}
\alias{STAT_RATING_PRIOR_ATTEMPTS_DEFAULT}
\alias{SKILL_PRIOR_ATTEMPTS_DEFAULT}
\title{Prior pseudo-attempts for Beta-Binomial efficiency stats}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
STAT_RATING_PRIOR_ATTEMPTS_DEFAULT

SKILL_PRIOR_ATTEMPTS_DEFAULT
}
\description{
Prior pseudo-attempts for Beta-Binomial efficiency stats
}
\keyword{internal}
