% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STAT_RATING_PRIOR_GAMES_DEFAULT}
\alias{STAT_RATING_PRIOR_GAMES_DEFAULT}
\alias{SKILL_PRIOR_GAMES_DEFAULT}
\title{Prior pseudo-games for Gamma-Poisson rate stats}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
STAT_RATING_PRIOR_GAMES_DEFAULT

SKILL_PRIOR_GAMES_DEFAULT
}
\description{
Prior pseudo-games for Gamma-Poisson rate stats
}
\keyword{internal}
