% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STAT_WEIGHT_LAMBDA}
\alias{STAT_WEIGHT_LAMBDA}
\title{L2 (ridge) regularization lambda for count-based stat weights}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
STAT_WEIGHT_LAMBDA
}
\description{
L2 (ridge) regularization lambda for count-based stat weights
}
\keyword{internal}
