% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{store_in_cache}
\alias{store_in_cache}
\title{Store data in cache}
\usage{
store_in_cache(cache_key, data)
}
\arguments{
\item{cache_key}{Character cache key}

\item{data}{Data to cache}
}
\value{
Invisible NULL
}
\description{
Store data in cache
}
\keyword{internal}
