% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{team_color_scale}
\alias{team_color_scale}
\title{Team colour scale (colour aesthetic)}
\usage{
team_color_scale(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot2 colour scale.
}
\description{
A convenience wrapper for \code{scale_colour_manual()} using \code{AFL_TEAM_COLORS}.
}
