% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams}
\alias{teams}
\title{Team information}
\format{
A data frame with X rows and Y variables:
\describe{
\item{team_id}{Unique identifier for each team}
\item{team_name}{Full name of the team}
\item{abbreviation}{Team abbreviation}
\item{conference}{Conference the team belongs to}
...
}
}
\source{
\url{http://example.com}
}
\usage{
teams
}
\description{
Team information
}
\keyword{datasets}
