% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{theme_torp}
\alias{theme_torp}
\title{torp ggplot2 theme}
\usage{
theme_torp(base_size = 12)
}
\arguments{
\item{base_size}{Base font size (default 12).}
}
\value{
A ggplot2 theme object.
}
\description{
A clean theme for torp visualizations based on \code{theme_minimal()}.
}
