% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{torp_df_total}
\alias{torp_df_total}
\title{Total torpedo data frame}
\format{
A data frame with X rows and Y variables:
\describe{
\item{variable1}{Description of variable1}
\item{variable2}{Description of variable2}
...
}
}
\source{
\url{http://example.com}
}
\usage{
torp_df_total
}
\description{
Total torpedo data frame
}
\keyword{datasets}
