% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{TORP_EPR_WEIGHT}
\alias{TORP_EPR_WEIGHT}
\title{Weight of EPR in TORP blend (0.5 = equal blend of EPR + PSR)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TORP_EPR_WEIGHT
}
\description{
Weight of EPR in TORP blend (0.5 = equal blend of EPR + PSR)
}
\keyword{internal}
