% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_api.R
\name{torp_replace_teams}
\alias{torp_replace_teams}
\title{Standardise AFL Team Names}
\usage{
torp_replace_teams(team)
}
\arguments{
\item{team}{Character vector of team names}
}
\value{
Character vector with standardised names
}
\description{
Maps team name variants (abbreviations, nicknames, Indigenous round names)
to canonical team names. Drop-in replacement for \code{fitzRoy::torp_replace_teams()}.
}
\examples{
torp_replace_teams("Adelaide Crows")
torp_replace_teams(c("GWS Giants", "Narrm", "WB"))

}
