% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_generic_data_quality}
\alias{validate_generic_data_quality}
\title{Validate Generic Data Quality}
\usage{
validate_generic_data_quality(data)
}
\arguments{
\item{data}{Generic dataframe}
}
\value{
List of generic data quality issues
}
\description{
Generic quality validation for any data type
}
\keyword{internal}
