% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{validate_rounds}
\alias{validate_rounds}
\title{Validate rounds}
\usage{
validate_rounds(rounds)
}
\arguments{
\item{rounds}{A numeric vector or TRUE}
}
\value{
Validated rounds vector
}
\description{
Validate rounds
}
\keyword{internal}
