% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{validate_seasons}
\alias{validate_seasons}
\title{Validate seasons}
\usage{
validate_seasons(seasons)
}
\arguments{
\item{seasons}{A numeric vector or TRUE}
}
\value{
A validated numeric vector of seasons
}
\description{
Validate seasons
}
\keyword{internal}
