% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALIDATION_HIGH_MISSING_THRESHOLD}
\alias{VALIDATION_HIGH_MISSING_THRESHOLD}
\title{High missing data threshold (proportion)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VALIDATION_HIGH_MISSING_THRESHOLD
}
\description{
High missing data threshold (proportion)
}
\keyword{internal}
