% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALIDATION_MIN_CALIBRATION_BIN_SIZE}
\alias{VALIDATION_MIN_CALIBRATION_BIN_SIZE}
\title{Minimum observations for calibration bins}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VALIDATION_MIN_CALIBRATION_BIN_SIZE
}
\description{
Minimum observations for calibration bins
}
\keyword{internal}
