% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versebus.R
\name{vb_write_manifest}
\alias{vb_write_manifest}
\title{Write a bus_manifest.json to \code{path}}
\usage{
vb_write_manifest(
  entries,
  tag,
  path,
  producer = vb_producer_info(),
  notes = ""
)
}
\arguments{
\item{entries}{list of vb_asset_entry() results (the FULL tag contents)}
}
\description{
Write a bus_manifest.json to \code{path}
}
\keyword{internal}
