% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{WP_CREDIT_DISP_SHARE}
\alias{WP_CREDIT_DISP_SHARE}
\title{Default disposer share of WPA in win probability credit assignment}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WP_CREDIT_DISP_SHARE
}
\description{
Default disposer share of WPA in win probability credit assignment
}
\keyword{internal}
