% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WP_FINAL_2_MINS_SECONDS}
\alias{WP_FINAL_2_MINS_SECONDS}
\title{Final 2 minutes threshold (seconds remaining)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WP_FINAL_2_MINS_SECONDS
}
\description{
Final 2 minutes threshold (seconds remaining)
}
\keyword{internal}
