% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WP_SCORE_SCALING}
\alias{WP_SCORE_SCALING}
\title{Score scaling factor for WP baseline model}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WP_SCORE_SCALING
}
\description{
Score scaling factor for WP baseline model
}
\keyword{internal}
