% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WP_TIME_SCALING}
\alias{WP_TIME_SCALING}
\title{Time scaling factor for WP baseline model}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
WP_TIME_SCALING
}
\description{
Time scaling factor for WP baseline model
}
\keyword{internal}
