% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_cache.R
\name{write_disk_cache}
\alias{write_disk_cache}
\title{Write Data to Disk Cache}
\usage{
write_disk_cache(url, data)
}
\arguments{
\item{url}{Character URL as cache key}

\item{data}{Data to cache}
}
\value{
Invisible NULL
}
\description{
Write Data to Disk Cache
}
\keyword{internal}
