% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{write_local_parquet}
\alias{write_local_parquet}
\title{Write a Parquet File to Local Storage}
\usage{
write_local_parquet(url, data)
}
\arguments{
\item{url}{Character URL used as the key (filename extracted from it)}

\item{data}{Data frame to write}
}
\value{
Invisible NULL
}
\description{
Write a Parquet File to Local Storage
}
\keyword{internal}
