% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{XRATING_HGA}
\alias{XRATING_HGA}
\title{xScore rating home-ground advantage, in points of expected margin}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
XRATING_HGA
}
\description{
xScore rating home-ground advantage, in points of expected margin
}
\keyword{internal}
