% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_match.R
\docType{data}
\name{XRATING_K}
\alias{XRATING_K}
\title{xScore rating update rate, in points of rating per point of prediction error}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
XRATING_K
}
\description{
xScore rating update rate, in points of rating per point of prediction error
}
\keyword{internal}
