% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{add_shot_geometry_variables}
\alias{add_shot_geometry_variables}
\title{Add Shot Geometry Variables}
\usage{
add_shot_geometry_variables(df, goal_width)
}
\arguments{
\item{df}{A dataframe containing shot data.}

\item{goal_width}{The width of the goal.}
}
\value{
A dataframe with additional shot geometry variables.
}
\description{
Add Shot Geometry Variables
}
\keyword{internal}
