% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_variables.R
\name{add_wp_vars}
\alias{add_wp_vars}
\title{Add Win Probability Variables}
\usage{
add_wp_vars(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with additional win probability-related variables.
}
\description{
This function adds win probability-related variables to the input dataframe.
}
