% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_GOAL_WIDTH}
\alias{AFL_GOAL_WIDTH}
\title{AFL goal post width in meters}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFL_GOAL_WIDTH
}
\description{
AFL goal post width in meters
}
\keyword{internal}
