% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_MAX_PERIODS}
\alias{AFL_MAX_PERIODS}
\title{Number of quarters in a standard AFL match}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFL_MAX_PERIODS
}
\description{
Number of quarters in a standard AFL match
}
\keyword{internal}
