% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{AFL_PLAY_GAME_SECONDS}
\alias{AFL_PLAY_GAME_SECONDS}
\title{Nominal playing time per AFL match in seconds (excludes stoppages)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
AFL_PLAY_GAME_SECONDS
}
\description{
Nominal playing time per AFL match in seconds (excludes stoppages)
}
\keyword{internal}
