% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_QUARTER_DURATION}
\alias{AFL_QUARTER_DURATION}
\title{Duration of AFL quarter in game seconds (approximately 20 minutes of game time)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFL_QUARTER_DURATION
}
\description{
Duration of AFL quarter in game seconds (approximately 20 minutes of game time)
}
\keyword{internal}
