% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_TEAM_ALIASES}
\alias{AFL_TEAM_ALIASES}
\title{Named vector mapping team name variants to canonical names}
\format{
An object of class \code{character} of length 93.
}
\usage{
AFL_TEAM_ALIASES
}
\description{
Maps abbreviations, full names, nicknames, Indigenous round names,
and other variants to the canonical short name used in \code{AFL_TEAMS$name}.
}
\keyword{datasets}
