% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AFL_TOTAL_GAME_SECONDS}
\alias{AFL_TOTAL_GAME_SECONDS}
\title{Total game duration in game seconds (4 quarters)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AFL_TOTAL_GAME_SECONDS
}
\description{
Total game duration in game seconds (4 quarters)
}
\keyword{internal}
