% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{calculate_player_stats}
\alias{calculate_player_stats}
\title{Calculate player statistics}
\usage{
calculate_player_stats(
  plyr_gm_df = NULL,
  match_ref,
  date_val,
  decay,
  loading,
  prior_games_recv,
  prior_games_disp
)
}
\arguments{
\item{plyr_gm_df}{Player game data frame}

\item{match_ref}{Match reference}

\item{date_val}{Date value}

\item{decay}{Decay factor}

\item{loading}{Loading factor}

\item{prior_games_recv}{Prior games for receiving}

\item{prior_games_disp}{Prior games for disposal}
}
\value{
A data frame with calculated player statistics
}
\description{
Calculate player statistics
}
