% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{calculate_torp_ratings}
\alias{calculate_torp_ratings}
\alias{torp_ratings}
\title{Get TORP ratings}
\usage{
calculate_torp_ratings(
  season_val = get_afl_season(type = "current"),
  round_val = get_afl_week(type = "next"),
  decay = RATING_DECAY_DEFAULT_DAYS,
  loading = 1.5,
  prior_games_recv = 4,
  prior_games_disp = 6,
  plyr_tm_df = NULL,
  player_game_data = NULL,
  spoil_multiplier = RATING_SPOIL_MULTIPLIER
)

torp_ratings(
  season_val = get_afl_season(type = "current"),
  round_val = get_afl_week(type = "next"),
  decay = RATING_DECAY_DEFAULT_DAYS,
  loading = 1.5,
  prior_games_recv = 4,
  prior_games_disp = 6,
  plyr_tm_df = NULL,
  player_game_data = NULL,
  spoil_multiplier = RATING_SPOIL_MULTIPLIER
)
}
\arguments{
\item{season_val}{The season to calculate ratings for. Default is the next season.}

\item{round_val}{The round to calculate ratings for. Default is the next round.}

\item{decay}{The decay factor for weighting games. Default is 365.}

\item{loading}{The loading factor for TORP calculations. Default is 1.5.}

\item{prior_games_recv}{The number of prior games to consider for receiving. Default is 4.}

\item{prior_games_disp}{The number of prior games to consider for disposal. Default is 6.}

\item{plyr_tm_df}{Optional pre-loaded player team data. If NULL, will load automatically.}

\item{player_game_data}{Optional pre-loaded player game data. If NULL, will load automatically.}

\item{spoil_multiplier}{Multiplier for spoil component. Default is \code{RATING_SPOIL_MULTIPLIER}.}
}
\value{
A data frame containing player TORP ratings.
}
\description{
This function calculates TORP (Total Overall Rating Points) for players based on their performance.

\code{torp_ratings()} is deprecated; use \code{calculate_torp_ratings()} instead.
}
