% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHAINS_CONTEST_TARGET_DESCS}
\alias{CHAINS_CONTEST_TARGET_DESCS}
\title{Chains descriptions for the contest target / kick result side
These appear in the row BEFORE the contest outcome (Spoil, Mark, etc.)
at the same x,y coordinates.}
\format{
An object of class \code{character} of length 2.
}
\usage{
CHAINS_CONTEST_TARGET_DESCS
}
\description{
Chains descriptions for the contest target / kick result side
These appear in the row BEFORE the contest outcome (Spoil, Mark, etc.)
at the same x,y coordinates.
}
\keyword{internal}
