% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CLOCK_DELTA_CAP}
\alias{CLOCK_DELTA_CAP}
\title{Maximum seconds between consecutive plays counted as playing time.
Gaps larger than this indicate unrecorded stoppages (injuries, reviews, etc.)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
CLOCK_DELTA_CAP
}
\description{
Maximum seconds between consecutive plays counted as playing time.
Gaps larger than this indicate unrecorded stoppages (injuries, reviews, etc.)
}
\keyword{internal}
