% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_afl.R
\docType{data}
\name{CLOCK_RESTART_EVENTS}
\alias{CLOCK_RESTART_EVENTS}
\title{Events that restart the play clock (chain \code{description} values)}
\format{
An object of class \code{character} of length 13.
}
\usage{
CLOCK_RESTART_EVENTS
}
\description{
Events that restart the play clock (chain \code{description} values)
}
\keyword{internal}
