% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CLOCK_STOPPAGE_TRIGGERS}
\alias{CLOCK_STOPPAGE_TRIGGERS}
\title{Descriptions after which the clock stops (scoring resets, ball out of play)}
\format{
An object of class \code{character} of length 6.
}
\usage{
CLOCK_STOPPAGE_TRIGGERS
}
\description{
Descriptions after which the clock stops (scoring resets, ball out of play)
}
\keyword{internal}
