% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{compare_models_statistical}
\alias{compare_models_statistical}
\title{Compare Multiple Models Statistically}
\usage{
compare_models_statistical(model_results, test_type = "simple")
}
\arguments{
\item{model_results}{List of model evaluation results}

\item{test_type}{Type of statistical test ("mcnemar", "delong")}
}
\value{
Dataframe with pairwise comparison results
}
\description{
Performs statistical comparison between multiple models using paired tests
}
