% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_models.R
\name{create_calibration_plot}
\alias{create_calibration_plot}
\title{Create Calibration Plot Data}
\usage{
create_calibration_plot(actual, predicted, n_bins = 10)
}
\arguments{
\item{actual}{Vector of actual binary outcomes}

\item{predicted}{Vector of predicted probabilities}

\item{n_bins}{Number of bins for calibration plot}
}
\value{
List with plot data and statistics
}
\description{
Creates data structure for plotting model calibration
}
\details{
A simplified wrapper around assess_model_calibration for creating calibration
plot data. For more comprehensive calibration analysis, use assess_model_calibration
directly followed by prepare_calibration_plot.
}
\seealso{
\code{\link{assess_model_calibration}}, \code{\link{prepare_calibration_plot}}
}
