% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{create_shot_model_matrix}
\alias{create_shot_model_matrix}
\title{Create Shot Model Matrix}
\usage{
create_shot_model_matrix(df)
}
\arguments{
\item{df}{A dataframe containing prepared shot data.}
}
\value{
A model matrix ready for shot modeling.
}
\description{
Create Shot Model Matrix
}
