% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_POS_ADJ_QUANTILE}
\alias{CREDIT_POS_ADJ_QUANTILE}
\title{Position-group quantile adjustment — legacy alias for backwards compatibility}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_POS_ADJ_QUANTILE
}
\description{
Position-group quantile adjustment — legacy alias for backwards compatibility
}
\keyword{internal}
