% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CREDIT_POS_ADJ_QUANTILE_HITOUT}
\alias{CREDIT_POS_ADJ_QUANTILE_HITOUT}
\title{Position-group quantile adjustment for hitout component}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CREDIT_POS_ADJ_QUANTILE_HITOUT
}
\description{
Position-group quantile adjustment for hitout component
}
\keyword{internal}
