% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decimal_hour}
\alias{decimal_hour}
\title{Convert datetime to decimal hours}
\usage{
decimal_hour(datetime)
}
\arguments{
\item{datetime}{A POSIXct datetime object.}
}
\value{
Numeric decimal hour.
}
\description{
Convert a POSIXct datetime to its decimal hour representation.
}
