% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ep_model}
\alias{ep_model}
\title{Expected Points Model for AFL Game Situations}
\format{
A GAM model object fitted using mgcv package
}
\source{
Created using data from AFL match tracking data via fitzRoy package
}
\usage{
ep_model
}
\description{
A GAM (Generalized Additive Model) fitted model object for predicting expected points
from various game situations based on field position, time, and game context.
}
\keyword{datasets}
