% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPR_PRIOR_RATE_HITOUT}
\alias{EPR_PRIOR_RATE_HITOUT}
\title{Prior rate for hitout component (shrinkage target per weighted game)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPR_PRIOR_RATE_HITOUT
}
\description{
Prior rate for hitout component (shrinkage target per weighted game)
}
\keyword{internal}
