% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epv_v3_duels.R
\name{epv3_aerial_out}
\alias{epv3_aerial_out}
\title{The outcome set the contest path should use, given the population setting}
\usage{
epv3_aerial_out(population = EPV3_CONTEST_POPULATION)
}
\description{
The outcome set the contest path should use, given the population setting
}
\keyword{internal}
