% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_data.R
\docType{data}
\name{EPV3_CONTEST_TARGET_DESCS}
\alias{EPV3_CONTEST_TARGET_DESCS}
\title{Chains descriptions v3 accepts as naming the beaten player in a duel}
\format{
An object of class \code{character} of length 1.
}
\usage{
EPV3_CONTEST_TARGET_DESCS
}
\description{
\verb{Contest Target} only. See \code{CHAINS_CONTEST_TARGET_DESCS} for why the
i50 annotation is excluded here and retained there.
}
\details{

}
\keyword{internal}
