% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV3_POINTS_SCALE}
\alias{EPV3_POINTS_SCALE}
\title{Per-channel points constants for v3 (1 unit = 1 point of margin)}
\format{
An object of class \code{numeric} of length 4.
}
\usage{
EPV3_POINTS_SCALE
}
\description{
v3's channels are separately calibrated, replacing the single global
\code{EPV_POINTS_SCALE}.
}
\details{

}
\keyword{internal}
