% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_data.R
\docType{data}
\name{EPV3_TARGET_FROM_I50}
\alias{EPV3_TARGET_FROM_I50}
\title{Also read the target off \verb{Kick Inside 50 Result}}
\format{
An object of class \code{logical} of length 1.
}
\usage{
EPV3_TARGET_FROM_I50
}
\description{
See \code{EPV3_CONTEST_TARGET_DESCS} for the measurement. \code{FALSE} keeps
the shipped behaviour of naming a beaten contestant only where a
\code{Contest Target} row happened to be logged.
}
\details{

}
\keyword{internal}
