% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_BOUNCE_WT}
\alias{EPV_BOUNCE_WT}
\title{Bounce weight in disposal credit}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_BOUNCE_WT
}
\description{
Bounce weight in disposal credit
}
\keyword{internal}
