% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_HITOUT_ADV_WT}
\alias{EPV_HITOUT_ADV_WT}
\title{Hitout to advantage weight}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_HITOUT_ADV_WT
}
\description{
Hitout to advantage weight
}
\keyword{internal}
