% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ratings.R
\docType{data}
\name{EPV_RECV_NEG_MULT}
\alias{EPV_RECV_NEG_MULT}
\title{Reception multiplier when defending (pos_team == -1)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_RECV_NEG_MULT
}
\description{
Reception multiplier when defending (pos_team == -1)
}
\keyword{internal}
