% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EPV_RUCK_CONTEST_WT}
\alias{EPV_RUCK_CONTEST_WT}
\title{Ruck contest weight (hitout component)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPV_RUCK_CONTEST_WT
}
\description{
Ruck contest weight (hitout component)
}
\keyword{internal}
