% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fixtures}
\alias{fixtures}
\title{Match schedule or fixture information}
\format{
A data frame with X rows and Y variables:
\describe{
\item{date}{Date of the match}
\item{home_team}{Home team name}
\item{away_team}{Away team name}
\item{venue}{Venue of the match}
...
}
}
\source{
\url{http://example.com}
}
\usage{
fixtures
}
\description{
Match schedule or fixture information
}
\keyword{datasets}
