% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_afl_season}
\alias{get_afl_season}
\title{Get AFL Season}
\usage{
get_afl_season(type = "current")
}
\arguments{
\item{type}{a TRUE/FALSE flag: if TRUE, returns the current year if March 1st or later. if FALSE, returns the current year if September 1st or later. Otherwise returns current year minus 1.}
}
\description{
A helper function to choose the most recent season available for a given dataset
}
