% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_afl_season}
\alias{get_afl_season}
\title{Get AFL Season}
\usage{
get_afl_season(type = "current")
}
\arguments{
\item{type}{A character string: "current" returns the current season, "next" returns the upcoming season.}
}
\value{
An integer representing the AFL season year.
}
\description{
A helper function to choose the most recent season available for a given dataset
}
