% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_from_cache}
\alias{get_from_cache}
\title{Get data from cache}
\usage{
get_from_cache(cache_key)
}
\arguments{
\item{cache_key}{Character cache key}
}
\value{
Cached data or NULL if not found
}
\description{
Get data from cache
}
\keyword{internal}
