% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_functions.R
\name{get_match_chains}
\alias{get_match_chains}
\title{Title}
\usage{
get_match_chains(season = lubridate::year(Sys.Date()), round = NA)
}
\arguments{
\item{round}{}
}
\description{
Title
}
\examples{
\dontrun{
get_match_chains()
}
}
