% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_functions.R
\name{get_players}
\alias{get_players}
\title{Get Players}
\usage{
get_players(use_api = FALSE)
}
\arguments{
\item{use_api}{Logical, whether to use the API (TRUE) or local database (FALSE, default).}
}
\value{
A dataframe containing player data.
}
\description{
Retrieves player data either from the API or from a local database.
}
