% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_proportion_through_day}
\alias{get_proportion_through_day}
\title{Get mode of a vector}
\usage{
get_proportion_through_day(datetime)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The mode (most frequent value) of the input vector.
}
\description{
This function returns the mode (most frequent value) of a vector.
}
