% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_functions.R
\name{get_season_games}
\alias{get_season_games}
\title{Get Season Games}
\usage{
get_season_games(season, rounds = 27)
}
\arguments{
\item{season}{The AFL season year (numeric).}

\item{rounds}{The number of rounds in the season (default: 27).}
}
\value{
A dataframe containing game data for the entire season.
}
\description{
Retrieves game data for an entire season.
}
