% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_functions.R
\name{get_token}
\alias{get_token}
\title{Get API Token}
\usage{
get_token()
}
\value{
A character string containing the API token.
}
\description{
Retrieves an authentication token for the AFL API.
}
\keyword{internal}
