% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_variables.R
\name{get_wp_preds}
\alias{get_wp_preds}
\title{Get Win Probability Predictions}
\usage{
get_wp_preds(df)
}
\arguments{
\item{df}{A dataframe containing play-by-play data.}
}
\value{
A dataframe with win probability predictions.
}
\description{
This function generates win probability predictions for the input data.
}
\keyword{internal}
