% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Vectorized Harmonic Mean of Two Numeric Vectors}
\usage{
harmonic_mean(x, y)
}
\arguments{
\item{x}{Numeric vector (e.g. home_shots).}

\item{y}{Numeric vector (e.g. away_shots).}
}
\value{
A numeric vector of harmonic means.
}
\description{
Computes the row-wise harmonic mean of two numeric vectors.
}
