% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R, R/utils.R
\name{is_installed}
\alias{is_installed}
\title{Check if a package is installed}
\usage{
is_installed(pkg)

is_installed(pkg)
}
\arguments{
\item{pkg}{Name of the package to check.}
}
\value{
Logical indicating if the package is installed

A logical value indicating whether the package is installed.
}
\description{
Check if a package is installed

Check if a package is installed
}
\keyword{internal}
