% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data.R
\name{is_locally_stored}
\alias{is_locally_stored}
\title{Check if a File Exists Locally}
\usage{
is_locally_stored(url)
}
\arguments{
\item{url}{Character URL to check for a local copy}
}
\value{
Logical
}
\description{
Check if a File Exists Locally
}
\keyword{internal}
