% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{load_chains}
\alias{load_chains}
\title{Load Chains Data}
\usage{
load_chains(seasons = get_afl_season(), rounds = get_afl_week())
}
\arguments{
\item{seasons}{A numeric vector of 4-digit years associated with given AFL seasons - defaults to latest season. If set to \code{TRUE}, returns all available data since 2021.}

\item{rounds}{A numeric vector associated with given AFL round - defaults to latest round. If set to \code{TRUE}, returns all available rounds in the given season range.}
}
\value{
A data frame containing chains data.
}
\description{
Loads chains data from the \href{https://github.com/peteowen1/torpdata}{torpdata repository}
}
\examples{
\donttest{
try({ # prevents cran errors
  load_chains(2021:2022)
})
}
}
\seealso{
\code{\link[=load_pbp]{load_pbp()}}, \code{\link[=load_xg]{load_xg()}}, \code{\link[=load_fixtures]{load_fixtures()}}
}
