% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{load_fixtures}
\alias{load_fixtures}
\title{Load Player Stats Data}
\usage{
load_fixtures(seasons = get_afl_season())
}
\arguments{
\item{seasons}{A numeric vector of 4-digit years associated with given AFL seasons - defaults to latest season. If set to \code{TRUE}, returns all available data since 2021.}
}
\value{
A data frame containing player stats data.
}
\description{
Loads player stats data from the \href{https://github.com/peteowen1/torpdata}{torpdata repository}
}
\examples{
\donttest{
try({ # prevents cran errors
  load_fixtures(2021:2022)
})
}
}
