% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{load_from_url}
\alias{load_from_url}
\title{Load parquet files from remote URLs}
\usage{
load_from_url(url, ..., seasons = TRUE, rounds = TRUE, peteowen1 = FALSE, use_disk_cache = TRUE)
}
\arguments{
\item{url}{A vector of URLs to load into memory. If more than one URL provided, will row-bind them.}

\item{...}{Named arguments that will be added as attributes to the data, e.g. \code{peteowen1_type} = "pbp"}

\item{seasons}{A numeric vector of years that will be used to filter the dataframe's \code{season} column. If \code{TRUE} (default), does not filter.}

\item{rounds}{A numeric vector of rounds that will be used to filter the dataframe's \code{round} column. If \code{TRUE} (default), does not filter.}

\item{peteowen1}{TRUE to add peteowen1_data classing and attributes.}
}
\value{
A data frame, possibly of type \code{peteowen1_data}
}
\description{
Load parquet files from remote URLs
}
