% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{load_pbp}
\alias{load_pbp}
\title{Load Play By Play}
\usage{
load_pbp(seasons = most_recent_season(), rounds = get_current_week())
}
\arguments{
\item{seasons}{A numeric vector of 4-digit years associated with given AFL seasons - defaults to latest season. If set to \code{TRUE}, returns all available data since 2021.}

\item{rounds}{A numeric vector associated with given AFL round - defaults to latest round. If set to \code{TRUE}, returns all available rounds in the given season range.}
}
\description{
Loads play by play seasons from the \href{https://github.com/peteowen1/torpdata}{torpdata repository}
}
\examples{
\donttest{
try({ # prevents cran errors
  load_pbp(2021:2022)
})
}

}
