% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{load_results}
\alias{load_results}
\title{Load AFL Match Results Data}
\usage{
load_results(seasons = get_afl_season())
}
\arguments{
\item{seasons}{A numeric vector of 4-digit years associated with given AFL seasons - defaults to latest season. If set to \code{TRUE}, returns all available data since 2021.}
}
\value{
A data frame containing AFL match results and final scores.
}
\description{
Loads AFL match results and scores from the \href{https://github.com/peteowen1/torpdata}{torpdata repository}
}
\examples{
\donttest{
try({ # prevents cran errors
  load_results(2021:2022)
})
}
}
\seealso{
\code{\link[=load_fixtures]{load_fixtures()}}, \code{\link[=load_predictions]{load_predictions()}}, \code{\link[=load_teams]{load_teams()}}
}
