% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{load_teams}
\alias{load_teams}
\title{Load AFL Team and Lineup Data}
\usage{
load_teams(seasons = get_afl_season())
}
\arguments{
\item{seasons}{A numeric vector of 4-digit years associated with given AFL seasons - defaults to latest season. If set to \code{TRUE}, returns all available data since 2021.}
}
\value{
A data frame containing AFL team and player lineup data.
}
\description{
Loads AFL team roster and lineup data from the \href{https://github.com/peteowen1/torpdata}{torpdata repository}
}
\examples{
\donttest{
try({ # prevents cran errors
  load_teams(2021:2022)
})
}
}
\seealso{
\code{\link[=load_fixtures]{load_fixtures()}}, \code{\link[=load_results]{load_results()}}, \code{\link[=load_player_details]{load_player_details()}}
}
