% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_xg_functions.R
\name{match_xgs}
\alias{match_xgs}
\title{Calculate xGs for AFL Matches}
\usage{
match_xgs(
  season = get_afl_season(),
  round = get_afl_week(),
  match_id = NA,
  quarter = 1:4
)
}
\arguments{
\item{season}{AFL season}

\item{round}{AFL round}

\item{match_id}{AFL match ID (not used in the current implementation)}

\item{quarter}{AFL match quarter}
}
\value{
A data frame with xG statistics for the specified matches
}
\description{
Calculate xGs for AFL Matches
}
