% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_xg_functions.R
\name{match_xgs}
\alias{match_xgs}
\title{match_xgs}
\usage{
match_xgs(season = get_afl_season(), round = get_afl_week(), match_id = NA)
}
\arguments{
\item{season}{AFL season}

\item{round}{AFL round}

\item{match_id}{AFL match ID}
}
\description{
match_xgs
}
