% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{parquet_from_url}
\alias{parquet_from_url}
\title{Load parquet file from a remote connection}
\usage{
parquet_from_url(url)
}
\arguments{
\item{url}{A character URL}
}
\value{
A data frame
}
\description{
Load parquet file from a remote connection
}
