% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{parquet_from_url_cached}
\alias{parquet_from_url_cached}
\title{Load parquet file from a remote connection with disk caching}
\usage{
parquet_from_url_cached(url, use_cache = TRUE, max_age_days = 7)
}
\arguments{
\item{url}{A character URL}

\item{use_cache}{Logical. If TRUE, use disk cache.}

\item{max_age_days}{Maximum age for cached files in days.}
}
\value{
A data frame
}
\description{
Load parquet file from a remote connection with disk caching
}
\keyword{internal}
