% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{player_game_ratings}
\alias{player_game_ratings}
\title{Get game ratings}
\usage{
player_game_ratings(
  season_val = get_afl_season(),
  round_num = get_afl_week(),
  matchid = FALSE,
  team = FALSE
)
}
\arguments{
\item{team}{}
}
\description{
Get game ratings
}
