% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{player_season_ratings}
\alias{player_season_ratings}
\title{Get season total ratings}
\usage{
player_season_ratings(season_val = get_afl_season(), round_num = NA)
}
\arguments{
\item{season_val}{}
}
\description{
Get season total ratings
}
