% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_shot_score}
\alias{player_shot_score}
\title{Player-specific shot scoring data}
\format{
A data frame with X rows and Y variables:
\describe{
\item{player_id}{Unique identifier for each player}
\item{shot_type}{Type of shot taken}
\item{score}{Score achieved}
...
}
}
\source{
\url{http://example.com}
}
\usage{
player_shot_score
}
\description{
Player-specific shot scoring data
}
\keyword{datasets}
