% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plyr_gm_df}
\alias{plyr_gm_df}
\title{Player game-level data}
\format{
A data frame with X rows and Y variables:
\describe{
\item{player_id}{Unique identifier for each player}
\item{game_id}{Unique identifier for each game}
\item{stat1}{Description of stat1}
\item{stat2}{Description of stat2}
...
}
}
\source{
\url{http://example.com}
}
\usage{
plyr_gm_df
}
\description{
Player game-level data
}
\keyword{datasets}
