% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{prepare_final_dataframe}
\alias{prepare_final_dataframe}
\title{Prepare final dataframe}
\usage{
prepare_final_dataframe(
  plyr_tm_df = NULL,
  plyr_gm_df = NULL,
  season_val,
  round_val
)
}
\arguments{
\item{plyr_tm_df}{Player team database}

\item{plyr_gm_df}{Player statistics dataframe}

\item{season_val}{Season value}

\item{round_val}{Round value}
}
\value{
A final dataframe with player ratings
}
\description{
Prepare final dataframe
}
