% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_cache.R
\name{read_disk_cache}
\alias{read_disk_cache}
\title{Read Data from Disk Cache}
\usage{
read_disk_cache(url)
}
\arguments{
\item{url}{Character URL to read cached data for}
}
\value{
Data frame if cache exists, NULL otherwise
}
\description{
Read Data from Disk Cache
}
\keyword{internal}
