% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{Game results data}
\format{
A data frame with X rows and Y variables:
\describe{
\item{game_id}{Unique identifier for each game}
\item{home_team}{Home team name}
\item{away_team}{Away team name}
\item{home_score}{Score of the home team}
\item{away_score}{Score of the away team}
...
}
}
\source{
\url{http://example.com}
}
\usage{
results
}
\description{
Game results data
}
\keyword{datasets}
