% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{safe_log_debug}
\alias{safe_log_debug}
\title{Safe Log Debug}
\usage{
safe_log_debug(message, ...)
}
\arguments{
\item{message}{Message to log}

\item{...}{Additional named parameters (ignored)}
}
\description{
Safe Log Debug
}
\keyword{internal}
