% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_logging.R
\name{safe_log_error}
\alias{safe_log_error}
\title{Safe Log Error}
\usage{
safe_log_error(message, ...)
}
\arguments{
\item{message}{Message to log}

\item{...}{Additional named parameters}
}
\description{
Safe Log Error
}
