% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_logging.R
\name{safe_log_info}
\alias{safe_log_info}
\title{Safe Logger Functions}
\usage{
safe_log_info(message, ...)
}
\arguments{
\item{message}{Message to log}

\item{...}{Additional named parameters}
}
\description{
These functions provide version-compatible logging that falls back to
base R messaging if logger package issues occur
}
\details{
Safe Log Info
}
