% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shot_player_df}
\alias{shot_player_df}
\title{Shot Player Reference Data}
\format{
A data frame with shot player references:
\describe{
\item{player_id_shot}{Unique identifier for the shooting player}
\item{player_name_shot}{Name of the shooting player}
}
}
\source{
AFL player reference data for shot analysis
}
\usage{
shot_player_df
}
\description{
Reference dataset containing player identifiers and names for
linking shot events to specific players.
}
\keyword{datasets}
