% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ratings.R
\name{torp_ratings}
\alias{torp_ratings}
\title{Get TORP ratings}
\usage{
torp_ratings(
  season_val = get_afl_season(type = "next"),
  round_val = get_afl_week(type = "next"),
  decay = 365,
  loading = 1.5,
  prior_games_recv = 3,
  prior_games_disp = 6
)
}
\arguments{
\item{prior_games}{}
}
\description{
Get TORP ratings
}
