% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl_api.R
\name{torp_replace_venues}
\alias{torp_replace_venues}
\title{Standardise AFL Venue Names}
\usage{
torp_replace_venues(venue)
}
\arguments{
\item{venue}{Character vector of venue names}
}
\value{
Character vector with standardised names
}
\description{
Maps venue name variants (sponsor names, old names) to canonical stable names.
Drop-in replacement for \code{fitzRoy::replace_venues()}.
}
\examples{
torp_replace_venues("Marvel Stadium")
torp_replace_venues(c("MCG", "Optus Stadium"))

}
