% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_torp_data.R
\name{validate_rounds}
\alias{validate_rounds}
\title{Validate seasons and rounds}
\usage{
validate_rounds(rounds)
}
\arguments{
\item{rounds}{A numeric vector or TRUE}

\item{seasons}{A numeric vector or TRUE}
}
\description{
Validate seasons and rounds
}
\keyword{internal}
