% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wp_model}
\alias{wp_model}
\title{Win probability model}
\format{
A model object or data frame with X rows and Y variables:
\describe{
\item{variable1}{Description of variable1}
\item{variable2}{Description of variable2}
...
}
}
\source{
\url{http://example.com}
}
\usage{
wp_model
}
\description{
Win probability model
}
\keyword{datasets}
